スペルディスペンサーのテストを分離
(cherry picked from commit 4d7122d2b7eb5efd7f7e611745a921313c424607)
