+++ src/main/java/jp/aquafactory/apprenticecodex/gametest/SpellDispenserGameTestScenarios.java
@@ -1,0 +1,2152 @@
+package jp.aquafactory.apprenticecodex.gametest;
+
+import com.mojang.authlib.GameProfile;
+import io.redspace.ironsspellbooks.api.events.CounterSpellEvent;
+import io.redspace.ironsspellbooks.api.events.SpellCooldownAddedEvent;
+import io.redspace.ironsspellbooks.api.events.SpellOnCastEvent;
+import io.redspace.ironsspellbooks.api.events.SpellPreCastEvent;
+import io.redspace.ironsspellbooks.api.item.UpgradeData;
+import io.redspace.ironsspellbooks.api.magic.MagicData;
+import io.redspace.ironsspellbooks.api.registry.SchoolRegistry;
+import io.redspace.ironsspellbooks.item.SpellSlotUpgradeItem;
+import io.redspace.ironsspellbooks.api.spells.IPresetSpellContainer;
+import io.redspace.ironsspellbooks.api.spells.AbstractSpell;
+import io.redspace.ironsspellbooks.api.spells.CastSource;
+import io.redspace.ironsspellbooks.api.spells.ISpellContainer;
+import io.redspace.ironsspellbooks.api.spells.SpellData;
+import io.redspace.ironsspellbooks.api.spells.SpellRarity;
+import io.redspace.ironsspellbooks.api.util.Utils;
+import io.redspace.ironsspellbooks.capabilities.magic.MagicManager;
+import io.redspace.ironsspellbooks.capabilities.magic.RecastInstance;
+import io.redspace.ironsspellbooks.effect.MagicMobEffect;
+import io.redspace.ironsspellbooks.entity.mobs.AntiMagicSusceptible;
+import io.redspace.ironsspellbooks.entity.spells.fire_breath.FireBreathProjectile;
+import io.redspace.ironsspellbooks.entity.spells.fireball.SmallMagicFireball;
+import io.redspace.ironsspellbooks.entity.spells.spectral_hammer.SpectralHammer;
+import io.redspace.ironsspellbooks.entity.spells.target_area.TargetedAreaEntity;
+import io.redspace.ironsspellbooks.spells.nature.TouchDigSpell;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
+import jp.aquafactory.apprenticecodex.ApprenticeCodex;
+import jp.aquafactory.apprenticecodex.block.arcanuminajar.ArcanumInAJarBlockEntity;
+import jp.aquafactory.apprenticecodex.block.atelierstation.AtelierStationBlockEntity;
+import jp.aquafactory.apprenticecodex.block.spellcalibrationbench.SpellCalibrationBenchMenu;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenser;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserBlockEntity;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserCastAnchorMode;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserCastHelper;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserCasterMode;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserManaFluidHelper;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserManaHelper;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserMenu;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserSpellProfile;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserSpellProfileManager;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserSpellValidator;
+import jp.aquafactory.apprenticecodex.block.spelldispenser.SpellDispenserVariant;
+import jp.aquafactory.apprenticecodex.block.spellcasterworkbench.SpellcasterWorkbenchMenu;
+import jp.aquafactory.apprenticecodex.capability.Capabilities;
+import jp.aquafactory.apprenticecodex.capability.codexspelldata.CodexSpellStateTypeRegister;
+import jp.aquafactory.apprenticecodex.compat.bettercombat.BetterCombatOffhandAttributeRescueCompat;
+import jp.aquafactory.apprenticecodex.config.ApprenticeCodexServerConfig;
+import jp.aquafactory.apprenticecodex.config.item.ArchivistsGrimoireServerConfig;
+import jp.aquafactory.apprenticecodex.config.item.SpellgunServerConfig;
+import jp.aquafactory.apprenticecodex.config.item.SpellThrowableCardServerConfig;
+import jp.aquafactory.apprenticecodex.enchantment.WisdomExperienceDropEvent;
+import jp.aquafactory.apprenticecodex.entity.spelldispenser.SpellDispenserAnchorEntity;
+import jp.aquafactory.apprenticecodex.damage.DamageTypes;
+import jp.aquafactory.apprenticecodex.config.item.SpellStainedRunicTabletServerConfig;
+import jp.aquafactory.apprenticecodex.datagen.DamageTypeTagGenerator;
+import jp.aquafactory.apprenticecodex.effect.CastingMoveSpeedAdjustment;
+import jp.aquafactory.apprenticecodex.enchantment.Enchantments;
+import jp.aquafactory.apprenticecodex.effect.PhalanxStance;
+import jp.aquafactory.apprenticecodex.event.ErrandMageVillagerTradesEvent;
+import jp.aquafactory.apprenticecodex.event.errandmage.ErrandMageTradeManager;
+import jp.aquafactory.apprenticecodex.event.ScrollcasterGauntletGrindstoneEvent;
+import jp.aquafactory.apprenticecodex.network.packet.SenseEvilHighlightsPacket;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerCastAnchorEntity;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerCastMode;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerCastOrigin;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerCastProfile;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerCastProfileManager;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerDirectionMode;
+import jp.aquafactory.apprenticecodex.remoteownercast.RemoteOwnerOriginMode;
+import jp.aquafactory.apprenticecodex.item.AbstractOffhandMagicItem;
+import jp.aquafactory.apprenticecodex.item.AbstractImbueShieldItem;
+import jp.aquafactory.apprenticecodex.item.AbstractRightClickMagicWeaponItem;
+import jp.aquafactory.apprenticecodex.item.AbstractSpellGunItem;
+import jp.aquafactory.apprenticecodex.item.AbstractSwingMagicItem;
+import jp.aquafactory.apprenticecodex.item.ChargedTwinBladeStaff;
+import jp.aquafactory.apprenticecodex.item.CircuitHeatStaff;
+import jp.aquafactory.apprenticecodex.item.CircuitHeatStaffCastEvent;
+import jp.aquafactory.apprenticecodex.item.CircuitHeatStaffRightClickItemEvent;
+import jp.aquafactory.apprenticecodex.item.CrystalBladedStaff;
+import jp.aquafactory.apprenticecodex.item.ElementalBow;
+import jp.aquafactory.apprenticecodex.item.FocusStaffbow;
+import jp.aquafactory.apprenticecodex.item.ammo.BowCastAmmoResolver;
+import jp.aquafactory.apprenticecodex.item.ItemManaBypassCastEvent;
+import jp.aquafactory.apprenticecodex.item.ManaBypassSpellItem;
+import jp.aquafactory.apprenticecodex.item.MithrilFreecastStaff;
+import jp.aquafactory.apprenticecodex.item.MultipurposeStaffrifle;
+import jp.aquafactory.apprenticecodex.item.RevolvercastStaff;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffAttackHandler;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffAttackProfile;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffAttackProfileManager;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffCastHelper;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffMobEffectHandler;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffMobEffectProfile;
+import jp.aquafactory.apprenticecodex.item.multicastechostaff.MulticastEchoStaffMobEffectProfileManager;
+import jp.aquafactory.apprenticecodex.item.multipurposestaffrifle.MultipurposeStaffrifleCastContext;
+import jp.aquafactory.apprenticecodex.item.multipurposestaffrifle.MultipurposeStaffrifleCastEvent;
+import jp.aquafactory.apprenticecodex.item.revolvercaststaff.RevolvercastStaffPendingAdvance;
+import jp.aquafactory.apprenticecodex.item.zenithstaff.ZenithStaffManaCostEvent;
+import jp.aquafactory.apprenticecodex.item.zenithstaff.ZenithStaffPowerHelper;
+import jp.aquafactory.apprenticecodex.item.NonDamageableAnvilMergeItem;
+import jp.aquafactory.apprenticecodex.item.RestrictedSpellImbuableItem;
+import jp.aquafactory.apprenticecodex.item.SmashcastScepter;
+import jp.aquafactory.apprenticecodex.item.smashcastscepter.SmashcastScepterAttackEvent;
+import jp.aquafactory.apprenticecodex.item.SpellGunCastEvent;
+import jp.aquafactory.apprenticecodex.item.SpellGunCastType;
+import jp.aquafactory.apprenticecodex.item.SpellcasterRoundItem;
+import jp.aquafactory.apprenticecodex.item.ScrollcasterGauntlet;
+import jp.aquafactory.apprenticecodex.item.ScrollcasterGauntletCastEvent;
+import jp.aquafactory.apprenticecodex.item.curios.autocastamulet.AutocastAmulet;
+import jp.aquafactory.apprenticecodex.item.curios.autocastamulet.AutocastAmuletAutoCastEvent;
+import jp.aquafactory.apprenticecodex.item.curios.autocastamulet.AutocastAmuletSpellListManager;
+import jp.aquafactory.apprenticecodex.item.curios.archivistsgrimoire.ArchivistsGrimoire;
+import jp.aquafactory.apprenticecodex.item.curios.craftsmansdelight.CraftsmansDelight;
+import jp.aquafactory.apprenticecodex.item.curios.craftsmansdelight.CraftsmansDelightCooldownReductionEvent;
+import jp.aquafactory.apprenticecodex.item.curios.craftsmansdelight.CraftsmansDelightManaCostDiscountEvent;
+import jp.aquafactory.apprenticecodex.item.curios.craftsmansdelight.CraftsmansDelightSpellSupport;
+import jp.aquafactory.apprenticecodex.item.curios.manashieldcharm.ManaShieldCharm;
+import jp.aquafactory.apprenticecodex.item.curios.spellstainedrunictablet.SpellStainedRunicTablet;
+import jp.aquafactory.apprenticecodex.item.curios.spellcasterquiver.SpellcasterQuiver;
+import jp.aquafactory.apprenticecodex.item.curios.spellcasterquiver.SpellcasterQuiverPickupEvent;
+import jp.aquafactory.apprenticecodex.item.flask.AlchemistsFlask;
+import jp.aquafactory.apprenticecodex.item.flask.AbstractPotionFlaskItem;
+import jp.aquafactory.apprenticecodex.item.flask.SpellcastersFlask;
+import jp.aquafactory.apprenticecodex.item.offhand.PhotonSiphon;
+import jp.aquafactory.apprenticecodex.item.shield.ReflectcastShield;
+import jp.aquafactory.apprenticecodex.item.spellthrowablecard.AbstractSpellThrowableCardItem;
+import jp.aquafactory.apprenticecodex.item.curios.CuriosSlotConstants;
+import jp.aquafactory.apprenticecodex.mixin.SinglePoolElementAccessor;
+import jp.aquafactory.apprenticecodex.mixin.StructureTemplatePoolAccessor;
+import jp.aquafactory.apprenticecodex.recipe.crafting.ExplorersCodexGuidebookTransferRecipe;
+import jp.aquafactory.apprenticecodex.capability.codexspelldata.spellstates.ManaShieldCharmState;
+import jp.aquafactory.apprenticecodex.capability.codexspelldata.spellstates.SearchBeaconState;
+import jp.aquafactory.apprenticecodex.spell.companiontrunk.CompanionTrunkEntity;
+import jp.aquafactory.apprenticecodex.spell.compoundphial.CompoundPhialProjectileEntity;
+import jp.aquafactory.apprenticecodex.spell.demicreatorwings.DemicreatorWings;
+import jp.aquafactory.apprenticecodex.spell.demicreatorwings.DemicreatorWingsManager;
+import jp.aquafactory.apprenticecodex.spell.divinepossession.DivinePossessionPowerHelper;
+import jp.aquafactory.apprenticecodex.spell.archermultiple.ArcherMultipleBowEntity;
+import jp.aquafactory.apprenticecodex.spell.assistwings.AssistWingsWingEntity;
+import jp.aquafactory.apprenticecodex.spell.automagnet.AutoMagnetFamiliarEntity;
+import jp.aquafactory.apprenticecodex.spell.automagnet.AutoMagnetFamiliarManager;
+import jp.aquafactory.apprenticecodex.spell.earthforge.EarthForge;
+import jp.aquafactory.apprenticecodex.spell.extract.ExtractPotionProjectileEntity;
+import jp.aquafactory.apprenticecodex.spell.flyswatter.FlySwatterProjectileEntity;
+import jp.aquafactory.apprenticecodex.spell.harvestmoon.HarvestMoon;
+import jp.aquafactory.apprenticecodex.spell.healingbloom.HealingBloom;
+import jp.aquafactory.apprenticecodex.spell.healingbloom.HealingBloomEntity;
+import jp.aquafactory.apprenticecodex.spell.healingbloom.HealingBloomLightBlockEntity;
+import jp.aquafactory.apprenticecodex.spell.ICraftsmansDelightAffectedSpell;
+import jp.aquafactory.apprenticecodex.spell.illuminatestellar.IlluminateStellarStarEntity;
+import jp.aquafactory.apprenticecodex.spell.inscribeice.InscribeIce;
+import jp.aquafactory.apprenticecodex.spell.inscribeice.InscribeIceBurst;
+import jp.aquafactory.apprenticecodex.spell.inscribeice.InscribeIceDaggerEntity;
+import jp.aquafactory.apprenticecodex.spell.magicspear.MagicSpearMissileEntity;
+import jp.aquafactory.apprenticecodex.spell.manaslash.ManaSlashProjectileEntity;
+import jp.aquafactory.apprenticecodex.spell.mysticshield.MysticShield;
+import jp.aquafactory.apprenticecodex.spell.mysticshield.MysticShieldDefenseEvent;
+import jp.aquafactory.apprenticecodex.spell.mysticshield.MysticShieldProjectileEntity;
+import jp.aquafactory.apprenticecodex.spell.mysticshield.MysticShieldShieldEntity;
+import jp.aquafactory.apprenticecodex.spell.personalshelf.PersonalShelf;
+import jp.aquafactory.apprenticecodex.spell.personalshelf.PersonalShelfChestBlockEntity;
+import jp.aquafactory.apprenticecodex.spell.phalanxcharge.PhalanxCounterSpellEvent;
+import jp.aquafactory.apprenticecodex.spell.precisionjack.PrecisionJackKnifeEntity;
+import jp.aquafactory.apprenticecodex.spell.senseevil.SenseEvil;
+import jp.aquafactory.apprenticecodex.spell.searchbeacon.SearchBeaconSearchService;
+import jp.aquafactory.apprenticecodex.spell.searchbeacon.SearchBeaconTargetList;
+import jp.aquafactory.apprenticecodex.spell.searchbeacon.SearchBeaconTargetManager;
+import jp.aquafactory.apprenticecodex.spell.skyedge.SkyEdgeProjectileEntity;
+import jp.aquafactory.apprenticecodex.spell.tinylumberjack.TinyLumberjackBlockClassifier;
+import jp.aquafactory.apprenticecodex.spell.tinylumberjack.TinyLumberjackJob;
+import jp.aquafactory.apprenticecodex.spell.uniteluna.UniteLunaMoonEntity;
+import jp.aquafactory.apprenticecodex.spell.worldflatter.WorldFlatterDrillEntity;
+import jp.aquafactory.apprenticecodex.item.armor.ApprenticeMageRobeItem;
+import jp.aquafactory.apprenticecodex.item.armor.ChromaticMagiaDressItem;
+import jp.aquafactory.apprenticecodex.item.armor.ChromaticMagiaDressStats;
+import jp.aquafactory.apprenticecodex.item.armor.ElementMaidenRobeItem;
+import jp.aquafactory.apprenticecodex.item.armor.ElementMaidenRobeSchoolPowerBonusEvents;
+import jp.aquafactory.apprenticecodex.item.armor.ElementMaidenRobeStats;
+import jp.aquafactory.apprenticecodex.item.swingstaff.AbstractSwingcastStaffItem;
+import jp.aquafactory.apprenticecodex.item.swingstaff.SwingcastCooldownMode;
+import jp.aquafactory.apprenticecodex.registry.ApprenticeAttributeRegistry;
+import jp.aquafactory.apprenticecodex.registry.BlockEntityRegistry;
+import jp.aquafactory.apprenticecodex.registry.BlockRegistry;
+import jp.aquafactory.apprenticecodex.registry.CreativeTabRegistry;
+import jp.aquafactory.apprenticecodex.registry.EffectRegistry;
+import jp.aquafactory.apprenticecodex.registry.EntityRegistry;
+import jp.aquafactory.apprenticecodex.registry.ItemRegistry;
+import jp.aquafactory.apprenticecodex.registry.LootConditionRegistry;
+import jp.aquafactory.apprenticecodex.registry.PoiTypeRegistry;
+import jp.aquafactory.apprenticecodex.registry.PotionRegistry;
+import jp.aquafactory.apprenticecodex.registry.RecipeRegistry;
+import jp.aquafactory.apprenticecodex.registry.SpellRegistry;
+import jp.aquafactory.apprenticecodex.item.armor.EnchantressRobeItem;
+import jp.aquafactory.apprenticecodex.item.armor.EnchantressRobeStats;
+import jp.aquafactory.apprenticecodex.item.armor.StealthRuneArmorItem;
+import jp.aquafactory.apprenticecodex.registry.TagRegistry;
+import jp.aquafactory.apprenticecodex.registry.VillagerProfessionRegistry;
+import jp.aquafactory.apprenticecodex.utility.BlockTools;
+import jp.aquafactory.apprenticecodex.utility.CombatTools;
+import jp.aquafactory.apprenticecodex.utility.InitialSpellContainerHelper;
+import jp.aquafactory.apprenticecodex.utility.MagicTools;
+import jp.aquafactory.apprenticecodex.utility.PresetSpellContainerStateHelper;
+import jp.aquafactory.apprenticecodex.utility.PotionContentsHelper;
+import jp.aquafactory.apprenticecodex.utility.BlockTools;
+import jp.aquafactory.apprenticecodex.utility.ErrandMageTradeHelper;
+import jp.aquafactory.apprenticecodex.utility.ProcessingRecipeDenylist;
+import jp.aquafactory.apprenticecodex.utility.RaycastTools;
+import jp.aquafactory.apprenticecodex.utility.RightClickSpellResolver;
+import jp.aquafactory.apprenticecodex.utility.SchoolAffinityRegistry;
+import jp.aquafactory.apprenticecodex.utility.SpellCalibrationImbueHelper;
+import jp.aquafactory.apprenticecodex.utility.ScrollcasterSchoolRuneResolver;
+import net.minecraft.ChatFormatting;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.core.NonNullList;
+import net.minecraft.core.GlobalPos;
+import net.minecraft.core.Holder;
+import net.minecraft.core.Registry;
+import net.minecraft.core.RegistryAccess;
+import net.minecraft.core.component.DataComponentPredicate;
+import net.minecraft.core.component.DataComponents;
+import net.minecraft.core.SectionPos;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.gametest.framework.GameTestHelper;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextColor;
+import net.minecraft.network.chat.contents.TranslatableContents;
+import net.minecraft.nbt.ListTag;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.EnchantmentTags;
+import net.minecraft.tags.PoiTypeTags;
+import net.minecraft.tags.DamageTypeTags;
+import net.minecraft.tags.TagKey;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.util.Mth;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.damagesource.DamageType;
+import net.minecraft.world.entity.decoration.ArmorStand;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.EquipmentSlotGroup;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.MoverType;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.ai.attributes.Attribute;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.memory.MemoryModuleType;
+import net.minecraft.world.entity.ai.village.poi.PoiTypes;
+import net.minecraft.world.entity.npc.VillagerData;
+import net.minecraft.world.entity.npc.VillagerProfession;
+import net.minecraft.world.entity.npc.VillagerTrades;
+import net.minecraft.world.SimpleContainer;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.ChestMenu;
+import net.minecraft.world.inventory.CraftingContainer;
+import net.minecraft.world.inventory.TransientCraftingContainer;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterials;
+import net.minecraft.world.item.ArrowItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.alchemy.Potions;
+import net.minecraft.world.item.component.CustomData;
+import net.minecraft.world.item.component.ItemAttributeModifiers;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraft.world.item.enchantment.ItemEnchantments;
+import net.minecraft.world.item.crafting.CraftingInput;
+import net.minecraft.world.item.crafting.RecipeManager;
+import net.minecraft.world.item.crafting.SingleRecipeInput;
+import net.minecraft.world.item.trading.ItemCost;
+import net.minecraft.world.item.trading.MerchantOffer;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.storage.loot.BuiltInLootTables;
+import net.minecraft.world.damagesource.CombatRules;
+import net.minecraft.world.level.block.AttachedStemBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.CropBlock;
+import net.minecraft.world.level.block.FlowerPotBlock;
+import net.minecraft.world.level.block.LayeredCauldronBlock;
+import net.minecraft.world.level.block.LiquidBlock;
+import net.minecraft.world.level.block.NetherWartBlock;
+import net.minecraft.world.level.block.StairBlock;
+import net.minecraft.world.level.block.entity.ChestBlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplate;
+import net.minecraft.world.level.block.entity.SpawnerBlockEntity;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import net.minecraft.world.level.levelgen.structure.pools.SinglePoolElement;
+import net.minecraft.world.level.levelgen.structure.pools.StructurePoolElement;
+import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructurePlaceSettings;
+import net.minecraft.world.level.storage.loot.LootContext;
+import net.minecraft.world.level.storage.loot.LootParams;
+import net.minecraft.world.level.storage.loot.functions.EnchantRandomlyFunction;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.Vec3;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.common.damagesource.DamageContainer;
+import net.neoforged.neoforge.common.util.FakePlayer;
+import net.neoforged.neoforge.event.entity.player.AttackEntityEvent;
+import net.neoforged.neoforge.event.entity.EntityJoinLevelEvent;
+import net.neoforged.neoforge.event.entity.player.ItemEntityPickupEvent;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
+import net.neoforged.neoforge.event.GrindstoneEvent;
+import net.neoforged.neoforge.event.entity.living.LivingDamageEvent;
+import net.neoforged.neoforge.event.entity.living.LivingExperienceDropEvent;
+import net.neoforged.neoforge.event.level.BlockDropsEvent;
+import net.neoforged.neoforge.event.ItemAttributeModifierEvent;
+import net.neoforged.neoforge.event.village.VillagerTradesEvent;
+import net.neoforged.neoforge.event.tick.PlayerTickEvent;
+import net.neoforged.neoforge.fluids.FluidStack;
+import net.neoforged.neoforge.fluids.capability.IFluidHandler;
+import net.neoforged.neoforge.items.ItemStackHandler;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.fml.ModList;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Comparator;
+import java.lang.reflect.Method;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
+
+final class SpellDispenserGameTestScenarios {
+
+    private static final String CREATE_GAMETEST_HOOKS_CLASS =
+            "jp.aquafactory.apprenticecodex.gametest.create.CreateGameTestHooks";
+    private static final TagKey<Item> CREATE_CONTRAPTION_CONTROLLED = TagKey.create(
+            Registries.ITEM,
+            ResourceLocation.fromNamespaceAndPath("create", "contraption_controlled")
+    );
+    private SpellDispenserGameTestScenarios() {
+    }
+
+    static void spellDispenserValidatorAcceptsSingleMagicMissileScroll(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(validation.isSupported(), "Spell Dispenser validator rejected a simple Magic Missile scroll");
+            helper.assertTrue(validation.spellData().getSpell() == io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get(),
+                    "Spell Dispenser validator resolved the wrong spell: " + validation.spellData().getSpell().getSpellResource());
+        });
+    }
+    static void spellDispenserValidatorAcceptsNonIronsNamespaceScroll(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(SpellRegistry.MAGE_LIGHT.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(validation.isSupported(), "Spell Dispenser validator rejected an apprenticecodex spell scroll");
+            helper.assertTrue(validation.spellData().getSpell() == SpellRegistry.MAGE_LIGHT.get(),
+                    "Spell Dispenser validator resolved the wrong non-Iron's spell: " + validation.spellData().getSpell().getSpellResource());
+        });
+    }
+    static void spellDispenserValidatorAcceptsLongScroll(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(SpellRegistry.COMPOUND_PHIAL.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(validation.isSupported(), "Spell Dispenser validator rejected a LONG scroll");
+            helper.assertTrue(validation.spellData().getSpell() == SpellRegistry.COMPOUND_PHIAL.get(),
+                    "Spell Dispenser validator resolved the wrong LONG spell: " + validation.spellData().getSpell().getSpellResource());
+        });
+    }
+    static void spellDispenserValidatorRejectsWhenServerDisabled(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    false,
+                    false,
+                    List.of(),
+                    1.0D
+            )) {
+                var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get());
+
+                var validation = SpellDispenserSpellValidator.validate(scrollStack);
+                helper.assertTrue(!validation.isSupported(), "Spell Dispenser validator accepted a scroll while server-disabled");
+                helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.SERVER_DISABLED,
+                        "Spell Dispenser validator returned the wrong failure reason while server-disabled: " + validation.failureReason());
+            }
+        });
+    }
+    static void spellDispenserValidatorRequiresServerAllowlist(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var magicMissile = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+            var magicMissileStack = createSpellScroll(magicMissile);
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    true,
+                    true,
+                    List.of(),
+                    1.0D
+            )) {
+                var validation = SpellDispenserSpellValidator.validate(magicMissileStack);
+                helper.assertTrue(!validation.isSupported(), "Spell Dispenser validator accepted a scroll with an empty server allowlist");
+                helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.NOT_ALLOWLISTED,
+                        "Spell Dispenser validator returned the wrong failure reason for an empty server allowlist: " + validation.failureReason());
+            }
+
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    true,
+                    true,
+                    List.of(magicMissile.getSpellResource().toString()),
+                    1.0D
+            )) {
+                var validation = SpellDispenserSpellValidator.validate(magicMissileStack);
+                helper.assertTrue(validation.isSupported(), "Spell Dispenser validator rejected an allowlisted profiled spell");
+            }
+
+            var unprofiledSpell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.RAY_OF_SIPHONING_SPELL.get();
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    true,
+                    true,
+                    List.of(unprofiledSpell.getSpellResource().toString()),
+                    1.0D
+            )) {
+                var validation = SpellDispenserSpellValidator.validate(createSpellScroll(unprofiledSpell));
+                helper.assertTrue(!validation.isSupported(), "Spell Dispenser server allowlist bypassed the profile requirement");
+                helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.NOT_PROFILED,
+                        "Spell Dispenser validator returned the wrong allowlist + unprofiled failure: " + validation.failureReason());
+            }
+        });
+    }
+    static void spellDispenserValidatorRejectsUnprofiledRayOfSiphoning(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.RAY_OF_SIPHONING_SPELL.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(!validation.isSupported(), "Spell Dispenser validator accepted Ray of Siphoning");
+            helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.NOT_PROFILED,
+                    "Spell Dispenser validator returned the wrong failure reason for Ray of Siphoning: " + validation.failureReason());
+        });
+    }
+    static void spellDispenserValidatorRejectsContinuousScroll(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(SpellRegistry.LONG_STRIDE.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(!validation.isSupported(), "Spell Dispenser validator accepted a CONTINUOUS scroll");
+            helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.NOT_PROFILED,
+                    "Spell Dispenser validator returned the wrong failure reason for profile-less CONTINUOUS scroll: " + validation.failureReason());
+        });
+    }
+    static void spellDispenserValidatorAcceptsProfiledContinuousScroll(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.FIRE_BREATH_SPELL.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(validation.isSupported(), "Spell Dispenser validator rejected a profiled CONTINUOUS scroll");
+            helper.assertTrue(validation.spellData().getSpell() == io.redspace.ironsspellbooks.api.registry.SpellRegistry.FIRE_BREATH_SPELL.get(),
+                    "Spell Dispenser validator resolved the wrong profiled CONTINUOUS spell: " + validation.spellData().getSpell().getSpellResource());
+        });
+    }
+    static void spellDispenserValidatorRejectsRecastScroll(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var scrollStack = createSpellScroll(SpellRegistry.HIGANBANA.get());
+
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(!validation.isSupported(), "Spell Dispenser validator accepted a recast spell scroll");
+            helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.HAS_RECAST,
+                    "Spell Dispenser validator returned the wrong failure reason for recast scroll: " + validation.failureReason());
+        });
+    }
+    static void spellDispenserValidatorRejectsNonScrollSpellContainer(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var nonScrollStack = createInitializedPresetStack(ItemRegistry.GRIMOIRE_MANIFEST.get());
+
+            helper.assertTrue(ISpellContainer.isSpellContainer(nonScrollStack), "Prepared non-scroll test stack is not a spell container");
+            var validation = SpellDispenserSpellValidator.validate(nonScrollStack);
+            helper.assertTrue(!validation.isSupported(), "Spell Dispenser validator accepted a non-scroll spell container");
+            helper.assertTrue(validation.failureReason() == SpellDispenserSpellValidator.FailureReason.NOT_SCROLL,
+                    "Spell Dispenser validator returned the wrong failure reason for non-scroll spell container: " + validation.failureReason());
+        });
+    }
+    static void spellDispenserCastHelperCleansUpProxy(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    (ServerLevel) level,
+                    castPos,
+                    Direction.NORTH,
+                    scrollStack,
+                    createSpellDispenserOwnerProfile("spell_dispenser_magic_missile_test")
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser cast helper failed to cast a Magic Missile scroll");
+            helper.assertTrue(castResult.reachedOnCast(), "Spell Dispenser Magic Missile cast did not mark that it reached onCast");
+            helper.assertTrue(castResult.cooldownTicks() == spell.getSpellCooldown(),
+                    "Spell Dispenser Magic Missile cast returned the wrong cooldown: " + castResult.cooldownTicks());
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Spell Dispenser proxy caster was left behind after Magic Missile cast");
+        });
+    }
+    static void spellDispenserCastHelperAllowsOwnerOptionalSpellWithoutOwnerProfile(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    (ServerLevel) level,
+                    castPos,
+                    Direction.NORTH,
+                    scrollStack,
+                    null
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser owner-optional cast failed without an owner profile");
+            helper.assertTrue(castResult.reachedOnCast(), "Spell Dispenser owner-optional cast did not reach onCast");
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Spell Dispenser owner-optional cast left proxy state behind");
+        });
+    }
+    static void spellDispenserCastHelperUsesNeutralLivingCasterProfileForMagicMissile(GameTestHelper helper) {
+        var level = (ServerLevel) helper.getLevel();
+        var castPos = new BlockPos(0, 1, 0);
+        var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+        var scrollStack = createSpellScroll(spell);
+        var neutralProfile = new SpellDispenserSpellProfile(
+                SpellDispenserCastAnchorMode.AUTO,
+                SpellDispenserCasterMode.NEUTRAL_LIVING,
+                0.0D,
+                0.0D,
+                0.0D,
+                0.0F,
+                0.0F,
+                false
+        );
+        var spawnedProjectiles = new ArrayList<io.redspace.ironsspellbooks.entity.spells.magic_missile.MagicMissileProjectile>();
+        java.util.function.Consumer<EntityJoinLevelEvent> projectileListener = event -> {
+            if (event.getLevel() == level
+                    && event.getEntity() instanceof io.redspace.ironsspellbooks.entity.spells.magic_missile.MagicMissileProjectile projectile) {
+                spawnedProjectiles.add(projectile);
+            }
+        };
+
+        NeoForge.EVENT_BUS.addListener(projectileListener);
+        try (var ignored = SpellDispenserSpellProfileManager.useProfilesForGameTest(Map.of(spell.getSpellResource(), neutralProfile))) {
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    level,
+                    castPos,
+                    Direction.NORTH,
+                    scrollStack,
+                    null
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser neutral living cast failed for Magic Missile");
+            helper.assertTrue(castResult.reachedOnCast(), "Spell Dispenser neutral living cast did not reach onCast");
+        } finally {
+            NeoForge.EVENT_BUS.unregister(projectileListener);
+        }
+
+        helper.assertTrue(!spawnedProjectiles.isEmpty(), "Spell Dispenser neutral living cast did not spawn Magic Missile");
+        var owner = spawnedProjectiles.get(0).getOwner();
+        helper.assertTrue(owner instanceof ArmorStand, "Spell Dispenser neutral living cast did not use ArmorStand owner: " + owner);
+        helper.assertTrue(owner.hasCustomName(), "Spell Dispenser neutral living caster did not set a display name");
+        helper.assertTrue(!owner.isCustomNameVisible(), "Spell Dispenser neutral living caster exposes its nameplate");
+        helper.assertTrue(!owner.isSilent(), "Spell Dispenser neutral living caster suppresses cast sound");
+        helper.assertTrue(!(owner instanceof Player), "Spell Dispenser neutral living cast used a Player owner: " + owner);
+        helper.assertTrue(!(owner instanceof FakePlayer), "Spell Dispenser neutral living cast used a FakePlayer owner: " + owner);
+        assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Spell Dispenser neutral living cast left proxy state behind");
+        helper.succeed();
+    }
+    static void spellDispenserCastHelperCompletesLongCastImmediately(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = SpellRegistry.COMPOUND_PHIAL.get();
+            var scrollStack = createSpellScroll(spell);
+
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    (ServerLevel) level,
+                    castPos,
+                    Direction.NORTH,
+                    scrollStack,
+                    createSpellDispenserOwnerProfile("spell_dispenser_long_test")
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser cast helper failed to cast a LONG Compound Phial scroll");
+            var expectedCooldown = spell.getSpellCooldown()
+                    + spell.getEffectiveCastTime(1, new FakePlayer((ServerLevel) level, createSpellDispenserOwnerProfile("spell_dispenser_long_probe")));
+            helper.assertTrue(castResult.cooldownTicks() == expectedCooldown,
+                    "Spell Dispenser LONG cast returned the wrong cooldown: " + castResult.cooldownTicks() + " / expected " + expectedCooldown);
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Spell Dispenser proxy caster was left behind after LONG cast");
+
+            var projectileBox = new AABB(castPos).inflate(5.0D);
+            var projectiles = level.getEntitiesOfClass(CompoundPhialProjectileEntity.class, projectileBox);
+            helper.assertTrue(!projectiles.isEmpty(), "Spell Dispenser LONG cast completed without spawning a Compound Phial projectile");
+        });
+    }
+    static void spellDispenserCastHelperTicksLongCastBeforeCompletion(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = SpellRegistry.PRECISION_JACK.get();
+            var scrollStack = createSpellScroll(spell);
+
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    (ServerLevel) level,
+                    castPos,
+                    Direction.NORTH,
+                    scrollStack,
+                    createSpellDispenserOwnerProfile("spell_dispenser_precision_jack_test")
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser cast helper failed to cast a Precision Jack scroll");
+
+            var knifeBox = new AABB(castPos).inflate(6.0D);
+            var knives = level.getEntitiesOfClass(PrecisionJackKnifeEntity.class, knifeBox);
+            helper.assertTrue(!knives.isEmpty(), "Spell Dispenser Precision Jack cast completed without spawning a knife");
+            helper.assertTrue(knives.stream().anyMatch(PrecisionJackKnifeEntity::isTrailActive),
+                    "Spell Dispenser LONG cast completed without calling Precision Jack's server cast tick");
+        });
+    }
+    static void spellDispenserPrecisionJackLowManaCleansUpKnife(GameTestHelper helper) {
+        var level = helper.getLevel();
+        var pos = new BlockPos(0, 1, 0);
+        var absolutePos = helper.absolutePos(pos);
+        helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+        var blockEntity = helper.getBlockEntity(pos);
+        helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+        var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+        var spell = SpellRegistry.PRECISION_JACK.get();
+        var requiredMana = spell.getManaCost(1);
+        spellDispenser.getInventory().setStackInSlot(0, createSpellScroll(spell));
+        spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_precision_jack_low_mana_test"));
+        spellDispenser.setCurrentMana(Math.max(0, requiredMana - 1));
+
+        var castResult = spellDispenser.tryActivate();
+        helper.assertTrue(!castResult.succeeded(), "Spell Dispenser activated Precision Jack with insufficient mana");
+        helper.assertTrue(castResult.insufficientMana(), "Spell Dispenser returned the wrong Precision Jack low-mana failure");
+
+        helper.runAtTickTime(20, () -> {
+            var knifeBox = new AABB(absolutePos).inflate(8.0D);
+            var knives = level.getEntitiesOfClass(PrecisionJackKnifeEntity.class, knifeBox);
+            helper.assertTrue(knives.isEmpty(), "Low-mana Precision Jack left knife entities behind: " + knives.size());
+            helper.succeed();
+        });
+    }
+    static void spellDispenserCastHelperStopsContinuousCastAtDurationCap(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.getSpell(
+                    ResourceLocation.fromNamespaceAndPath("irons_spellbooks", "blaze_storm")
+            );
+            var scrollStack = createSpellScroll(spell);
+            var validation = createSpellDispenserValidation(scrollStack, spell);
+
+            var startResult = SpellDispenserCastHelper.tryStartContinuousCast(
+                    level,
+                    castPos,
+                    Direction.NORTH,
+                    validation,
+                    scrollStack,
+                    createSpellDispenserOwnerProfile("spell_dispenser_continuous_test")
+            );
+            helper.assertTrue(startResult.result().succeeded(), "Spell Dispenser cast helper failed to start a CONTINUOUS Blaze Storm scroll");
+            helper.assertTrue(startResult.session() != null, "Spell Dispenser cast helper did not return a CONTINUOUS session");
+
+            var session = startResult.session();
+            var maxTicks = session.magicData().getCastDuration();
+            for (var tick = 0; tick <= maxTicks + 1 && !session.isFinished(); tick++) {
+                SpellDispenserCastHelper.tickContinuousCast(level, session);
+            }
+
+            helper.assertTrue(session.isFinished(), "Spell Dispenser CONTINUOUS cast did not stop by its cast duration cap");
+            helper.assertTrue(session.consumeFinishedCooldownTicks() == spell.getSpellCooldown(),
+                    "Spell Dispenser CONTINUOUS cast did not record the expected cooldown after completion");
+            var fireballBox = new AABB(castPos).inflate(16.0D);
+            var fireballs = level.getEntitiesOfClass(SmallMagicFireball.class, fireballBox);
+            helper.assertTrue(!fireballs.isEmpty(), "Spell Dispenser CONTINUOUS cast completed without spawning Blaze Storm fireballs");
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Spell Dispenser proxy caster was left behind after CONTINUOUS cast");
+        });
+    }
+    static void spellDispenserContinuousBreathBindsProjectileToTrackedAnchor(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.FIRE_BREATH_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+            var validation = SpellDispenserSpellValidator.validate(scrollStack);
+            helper.assertTrue(validation.isSupported(), "Spell Dispenser validator rejected Fire Breath profile");
+
+            var startResult = SpellDispenserCastHelper.tryStartContinuousCast(
+                    level,
+                    castPos,
+                    Direction.NORTH,
+                    validation,
+                    scrollStack,
+                    createSpellDispenserOwnerProfile("spell_dispenser_fire_breath_test")
+            );
+            helper.assertTrue(startResult.result().succeeded(), "Spell Dispenser cast helper failed to start Fire Breath");
+            helper.assertTrue(startResult.session() != null, "Spell Dispenser cast helper did not return a Fire Breath session");
+            helper.assertTrue(startResult.session().spellCaster() instanceof SpellDispenserAnchorEntity,
+                    "Spell Dispenser Fire Breath did not select a tracked living caster from its explicit profile: " + startResult.session().spellCaster());
+
+            var session = startResult.session();
+            for (var tick = 0; tick < 20; tick++) {
+                SpellDispenserCastHelper.tickContinuousCast(level, session);
+            }
+
+            var coneProjectiles = level.getEntitiesOfClass(FireBreathProjectile.class, new AABB(castPos).inflate(16.0D));
+            helper.assertTrue(!coneProjectiles.isEmpty(), "Spell Dispenser Fire Breath never spawned its cone projectile");
+            helper.assertTrue(coneProjectiles.get(0).getOwner() instanceof SpellDispenserAnchorEntity,
+                    "Spell Dispenser Fire Breath did not rebind its cone owner to a tracked anchor: " + coneProjectiles.get(0).getOwner());
+
+            SpellDispenserCastHelper.finishContinuousCast(level, session, true);
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Spell Dispenser tracked anchor was left behind after Fire Breath cleanup");
+        });
+    }
+    static void spellDispenserBlockEntityStopsContinuousCastWhenSignalTurnsOff(GameTestHelper helper) {
+        var level = (ServerLevel) helper.getLevel();
+        var pos = new BlockPos(0, 1, 0);
+        helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+        helper.setBlock(pos, helper.getBlockState(pos).setValue(SpellDispenser.TRIGGERED, true));
+
+        var blockEntity = helper.getBlockEntity(pos);
+        helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+        var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+        var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.getSpell(
+                ResourceLocation.fromNamespaceAndPath("irons_spellbooks", "blaze_storm")
+        );
+        var scrollStack = createSpellScroll(spell);
+        spellDispenser.getInventory().setStackInSlot(0, scrollStack);
+        spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_signal_stop_test"));
+
+        helper.runAtTickTime(1, () -> {
+            var startResult = SpellDispenserCastHelper.tryStartContinuousCast(
+                    level,
+                    pos,
+                    Direction.NORTH,
+                    createSpellDispenserValidation(scrollStack, spell),
+                    scrollStack,
+                    spellDispenser.getOwnerProfile()
+            );
+            helper.assertTrue(startResult.result().succeeded(), "Spell Dispenser block entity failed to prepare a CONTINUOUS session");
+            helper.assertTrue(startResult.session() != null, "Spell Dispenser block entity did not receive a powered CONTINUOUS session");
+            spellDispenser.startContinuousCast(startResult.session());
+        });
+        helper.runAtTickTime(2, () -> {
+            SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            helper.assertTrue(spellDispenser.hasActiveContinuousCast(), "Spell Dispenser CONTINUOUS cast stopped immediately while powered");
+            helper.setBlock(pos, helper.getBlockState(pos).setValue(SpellDispenser.TRIGGERED, false));
+        });
+        helper.runAtTickTime(3, () -> {
+            SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            helper.assertFalse(spellDispenser.hasActiveContinuousCast(), "Spell Dispenser CONTINUOUS cast did not stop after redstone signal loss");
+            assertNoSpellDispenserProxy(helper, pos, scrollStack, "Spell Dispenser proxy caster was left behind after signal loss");
+            helper.succeed();
+        });
+    }
+    static void spellDispenserBlockEntityDoesNotRestartContinuousCastWhileSignalStaysOn(GameTestHelper helper) {
+        var level = (ServerLevel) helper.getLevel();
+        var pos = new BlockPos(0, 1, 0);
+        helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+        helper.setBlock(pos, helper.getBlockState(pos).setValue(SpellDispenser.TRIGGERED, true));
+
+        var blockEntity = helper.getBlockEntity(pos);
+        helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+        var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+        var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.getSpell(
+                ResourceLocation.fromNamespaceAndPath("irons_spellbooks", "blaze_storm")
+        );
+        var scrollStack = createSpellScroll(spell);
+        spellDispenser.getInventory().setStackInSlot(0, scrollStack);
+        spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_signal_hold_test"));
+
+        helper.runAtTickTime(1, () -> {
+            var startResult = SpellDispenserCastHelper.tryStartContinuousCast(
+                    level,
+                    pos,
+                    Direction.NORTH,
+                    createSpellDispenserValidation(scrollStack, spell),
+                    scrollStack,
+                    spellDispenser.getOwnerProfile()
+            );
+            helper.assertTrue(startResult.result().succeeded(), "Spell Dispenser block entity failed to prepare a held-signal CONTINUOUS session");
+            helper.assertTrue(startResult.session() != null, "Spell Dispenser block entity did not receive a held-signal CONTINUOUS session");
+            spellDispenser.startContinuousCast(startResult.session());
+
+            var maxTicks = startResult.session().magicData().getCastDuration();
+            for (var tick = 0; tick <= maxTicks + 20; tick++) {
+                SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            }
+
+            helper.assertFalse(spellDispenser.hasActiveContinuousCast(),
+                    "Spell Dispenser CONTINUOUS cast restarted even though the redstone signal never toggled off");
+            assertNoSpellDispenserProxy(helper, pos, scrollStack, "Spell Dispenser proxy caster was left behind after held-signal completion");
+            helper.succeed();
+        });
+    }
+    static void spellDispenserBlockEntityMarksContinuousResetRequirementAfterHeldCompletion(GameTestHelper helper) {
+        var level = (ServerLevel) helper.getLevel();
+        var pos = new BlockPos(0, 1, 0);
+        helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+        helper.setBlock(pos, helper.getBlockState(pos).setValue(SpellDispenser.TRIGGERED, true));
+
+        var blockEntity = helper.getBlockEntity(pos);
+        helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+        var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+        var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.getSpell(
+                ResourceLocation.fromNamespaceAndPath("irons_spellbooks", "blaze_storm")
+        );
+        var scrollStack = createSpellScroll(spell);
+        spellDispenser.getInventory().setStackInSlot(0, scrollStack);
+        spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_reset_required_test"));
+
+        helper.runAtTickTime(1, () -> {
+            var startResult = SpellDispenserCastHelper.tryStartContinuousCast(
+                    level,
+                    pos,
+                    Direction.NORTH,
+                    createSpellDispenserValidation(scrollStack, spell),
+                    scrollStack,
+                    spellDispenser.getOwnerProfile()
+            );
+            helper.assertTrue(startResult.result().succeeded(), "Spell Dispenser block entity failed to prepare a reset-required CONTINUOUS session");
+            helper.assertTrue(startResult.session() != null, "Spell Dispenser block entity did not receive a reset-required CONTINUOUS session");
+            spellDispenser.startContinuousCast(startResult.session());
+
+            var maxTicks = startResult.session().magicData().getCastDuration();
+            for (var tick = 0; tick <= maxTicks + 20; tick++) {
+                SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            }
+
+            helper.assertFalse(spellDispenser.hasActiveContinuousCast(),
+                    "Spell Dispenser CONTINUOUS cast stayed active after its held-signal duration cap");
+            helper.assertTrue(spellDispenser.requiresContinuousReset(),
+                    "Spell Dispenser did not record that CONTINUOUS needs a signal reset after held completion");
+            helper.assertTrue(spellDispenser.isCoolingDown(),
+                    "Spell Dispenser did not enter cooldown after a held CONTINUOUS cast completed");
+            assertNoSpellDispenserProxy(helper, pos, scrollStack, "Spell Dispenser proxy caster was left behind after held completion reset state");
+            helper.succeed();
+        });
+    }
+    static void spellDispenserBlockEntityStartsAndClearsInstantCooldown(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+            spellDispenser.getInventory().setStackInSlot(0, scrollStack);
+            spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_cooldown_test"));
+
+            var firstResult = spellDispenser.tryActivate();
+            helper.assertTrue(firstResult.succeeded(), "Spell Dispenser failed to activate an INSTANT spell for cooldown test");
+            helper.assertTrue(spellDispenser.getRemainingCooldownTicks() == spell.getSpellCooldown(),
+                    "Spell Dispenser stored the wrong cooldown after an INSTANT cast: " + spellDispenser.getRemainingCooldownTicks());
+
+            var blockedResult = spellDispenser.tryActivate();
+            helper.assertTrue(!blockedResult.succeeded() && !blockedResult.reachedOnCast(),
+                    "Spell Dispenser did not reject activation while cooling down");
+
+            for (var tick = 0; tick < spell.getSpellCooldown(); tick++) {
+                SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            }
+
+            helper.assertFalse(spellDispenser.isCoolingDown(), "Spell Dispenser cooldown did not expire after the expected number of ticks");
+        });
+    }
+    static void spellDispenserBlockEntityReportsServerDisabledActivation(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    false,
+                    false,
+                    List.of(),
+                    1.0D
+            )) {
+                var pos = new BlockPos(0, 1, 0);
+                helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+                var blockEntity = helper.getBlockEntity(pos);
+                helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+                var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+                spellDispenser.getInventory().setStackInSlot(
+                        0,
+                        createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get())
+                );
+
+                var result = spellDispenser.tryActivate();
+                helper.assertTrue(!result.succeeded(), "Spell Dispenser activated while disabled by server config");
+                helper.assertTrue(result.failureType() == SpellDispenserCastHelper.FailureType.SERVER_DISABLED,
+                        "Spell Dispenser returned the wrong server-disabled activation failure: " + result.failureType());
+                helper.assertFalse(spellDispenser.isCoolingDown(),
+                        "Spell Dispenser entered cooldown despite server-disabled activation");
+            }
+        });
+    }
+    static void spellDispenserBlockEntityAppliesCooldownMultiplier(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    true,
+                    false,
+                    List.of(),
+                    0.1D
+            )) {
+                var pos = new BlockPos(0, 1, 0);
+                helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+                var blockEntity = helper.getBlockEntity(pos);
+                helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+                var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+                var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+                spellDispenser.getInventory().setStackInSlot(0, createSpellScroll(spell));
+                spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_cooldown_multiplier_test"));
+
+                var result = spellDispenser.tryActivate();
+                var expectedCooldown = Math.max(1, (int) Math.ceil(spell.getSpellCooldown() * 0.1D));
+                helper.assertTrue(result.succeeded(), "Spell Dispenser failed to activate for cooldown multiplier test");
+                helper.assertTrue(result.cooldownTicks() == expectedCooldown,
+                        "Spell Dispenser returned the wrong scaled cooldown: " + result.cooldownTicks() + " / expected " + expectedCooldown);
+                helper.assertTrue(spellDispenser.getRemainingCooldownTicks() == expectedCooldown,
+                        "Spell Dispenser stored the wrong scaled cooldown: " + spellDispenser.getRemainingCooldownTicks());
+            }
+        });
+    }
+    static void spellDispenserAutomationOnlyAcceptsManaContainers(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var blockEntity = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            var itemHandler = blockEntity.getItemHandler(Direction.UP);
+            var manaFlask = createFilledSpellcastersFlask(
+                    helper.getLevel().registryAccess(),
+                    createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get()),
+                    1,
+                    0
+            );
+            var nonManaPotion = PotionContentsHelper.createPotionStack(Items.POTION, net.minecraft.world.item.alchemy.Potions.HEALING.value());
+            var nonManaFlask = createFilledSpellcastersFlask(helper.getLevel().registryAccess(), nonManaPotion, 1, 0);
+
+            helper.assertTrue(itemHandler != null, "Spell Dispenser item capability was not exposed for flask slot validation");
+            helper.assertTrue(
+                    itemHandler != null && itemHandler.isItemValid(
+                            SpellDispenserBlockEntity.FLASK_SLOT_START,
+                            createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get())
+                    ),
+                    "Spell Dispenser flask slot rejected a mana potion"
+            );
+            helper.assertTrue(
+                    itemHandler != null && itemHandler.isItemValid(
+                            SpellDispenserBlockEntity.FLASK_SLOT_START,
+                            manaFlask
+                    ),
+                    "Spell Dispenser flask slot rejected a mana flask"
+            );
+            helper.assertTrue(
+                    itemHandler != null && !itemHandler.isItemValid(
+                            SpellDispenserBlockEntity.FLASK_SLOT_START,
+                            new ItemStack(Items.GLASS_BOTTLE)
+                    ),
+                    "Spell Dispenser flask slot accepted a glass bottle from automation"
+            );
+            helper.assertTrue(
+                    itemHandler != null && !itemHandler.isItemValid(
+                            SpellDispenserBlockEntity.FLASK_SLOT_START,
+                            nonManaPotion
+                    ),
+                    "Spell Dispenser flask slot accepted a non-mana potion"
+            );
+            helper.assertTrue(
+                    itemHandler != null && !itemHandler.isItemValid(
+                            SpellDispenserBlockEntity.FLASK_SLOT_START,
+                            nonManaFlask
+                    ),
+                    "Spell Dispenser flask slot accepted a non-mana flask"
+            );
+            helper.assertTrue(
+                    itemHandler != null && !itemHandler.isItemValid(
+                            SpellDispenserBlockEntity.FLASK_SLOT_START,
+                            new ItemStack(Items.STICK)
+                    ),
+                    "Spell Dispenser flask slot accepted an unrelated item"
+            );
+        });
+    }
+    static void spellDispenserFluidAcceptsOnlyRegularManaPotions(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var blockEntity = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            var fluidHandler = blockEntity.getFluidHandler(Direction.UP);
+            var partialManaFluid = createIronsManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get(), 100);
+            var remainingPartialManaFluid = createIronsManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get(), 150);
+            var strongerManaFluid = createIronsManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_TWO.get(), 250);
+            var splashPotion = PotionContentsHelper.createPotionStack(
+                    Items.SPLASH_POTION,
+                    io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get()
+            );
+            var splashManaFluid = io.redspace.ironsspellbooks.fluids.PotionFluid.from(splashPotion);
+            var healingFluid = createIronsManaPotionFluid(net.minecraft.world.item.alchemy.Potions.HEALING.value(), 250);
+
+            blockEntity.setCurrentMana(950);
+            helper.assertTrue(fluidHandler != null, "Spell Dispenser fluid capability was not exposed");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(healingFluid, IFluidHandler.FluidAction.EXECUTE) == 0,
+                    "Spell Dispenser accepted a non-mana potion fluid");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(splashManaFluid, IFluidHandler.FluidAction.EXECUTE) == 0,
+                    "Spell Dispenser accepted a splash mana potion fluid");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(partialManaFluid, IFluidHandler.FluidAction.EXECUTE) == 100,
+                    "Spell Dispenser rejected a partial regular mana potion fluid");
+            helper.assertTrue(blockEntity.getStoredManaPotionFluid().getAmount() == 100,
+                    "Spell Dispenser did not retain a partial mana potion fluid");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(strongerManaFluid, IFluidHandler.FluidAction.EXECUTE) == 0,
+                    "Spell Dispenser accepted a different mana potion fluid into the same tank");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(remainingPartialManaFluid, IFluidHandler.FluidAction.EXECUTE) == 150,
+                    "Spell Dispenser rejected matching mana potion fluid after a partial fill");
+            helper.assertTrue(blockEntity.getStoredManaPotionFluid().getAmount() == 250,
+                    "Spell Dispenser consumed a potion fluid that should not fit in remaining mana capacity");
+        });
+    }
+    static void spellDispenserFluidConsumesPotionDoseImmediately(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var blockEntity = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            var fluidHandler = blockEntity.getFluidHandler(null);
+            var manaFluid = createIronsManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get(), 250);
+
+            blockEntity.setCurrentMana(850);
+            helper.assertTrue(fluidHandler != null, "Spell Dispenser fluid capability was not exposed");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(manaFluid, IFluidHandler.FluidAction.EXECUTE) == 250,
+                    "Spell Dispenser did not accept a regular mana potion fluid");
+            helper.assertTrue(blockEntity.getCurrentMana() == 925,
+                    "Spell Dispenser did not immediately recover mana from potion fluid: " + blockEntity.getCurrentMana());
+            helper.assertTrue(blockEntity.getStoredManaPotionFluid().isEmpty(),
+                    "Spell Dispenser left consumed mana potion fluid in the tank");
+        });
+    }
+    static void spellDispenserFluidDrainsUnconsumedPotionOnly(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var blockEntity = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            var fluidHandler = blockEntity.getFluidHandler(Direction.NORTH);
+            var manaFluid = createIronsManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get(), 500);
+
+            blockEntity.setCurrentMana(950);
+            helper.assertTrue(fluidHandler != null, "Spell Dispenser fluid capability was not exposed");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(manaFluid, IFluidHandler.FluidAction.EXECUTE) == 500,
+                    "Spell Dispenser did not accept unconsumed mana potion fluid");
+
+            var drained = fluidHandler == null ? FluidStack.EMPTY : fluidHandler.drain(500, IFluidHandler.FluidAction.EXECUTE);
+            helper.assertTrue(drained.getAmount() == 500,
+                    "Spell Dispenser did not drain all unconsumed potion fluid: " + drained.getAmount());
+            helper.assertTrue(blockEntity.getCurrentMana() == 950,
+                    "Spell Dispenser converted internal mana into drained potion fluid");
+            helper.assertTrue(fluidHandler == null || fluidHandler.drain(250, IFluidHandler.FluidAction.EXECUTE).isEmpty(),
+                    "Spell Dispenser drained potion fluid after the tank was empty");
+        });
+    }
+    static void spellDispenserFluidPersistsThroughNbt(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var original = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            var fluidHandler = original.getFluidHandler(Direction.SOUTH);
+            var manaFluid = createIronsManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get(), 250);
+
+            original.setCurrentMana(950);
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(manaFluid, IFluidHandler.FluidAction.EXECUTE) == 250,
+                    "Spell Dispenser did not accept mana potion fluid before NBT round-trip");
+
+            var restored = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            restored.loadWithComponents(original.getUpdateTag(helper.getLevel().registryAccess()), helper.getLevel().registryAccess());
+            helper.assertTrue(restored.getStoredManaPotionFluid().getAmount() == 250,
+                    "Spell Dispenser did not restore stored mana potion fluid from NBT");
+            helper.assertTrue(SpellDispenserManaFluidHelper.isSameFluidAndTags(restored.getStoredManaPotionFluid(), manaFluid),
+                    "Spell Dispenser changed stored potion fluid identity during NBT round-trip");
+        });
+    }
+    static void spellDispenserFluidAcceptsCreateManaPotion(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var blockEntity = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            var fluidHandler = blockEntity.getFluidHandler(Direction.WEST);
+            var manaFluid = createCreateManaPotionFluid(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get(), 250);
+
+            blockEntity.setCurrentMana(850);
+            helper.assertTrue(!manaFluid.isEmpty(), "Create potion fluid was not registered");
+            helper.assertTrue(fluidHandler != null && fluidHandler.fill(manaFluid, IFluidHandler.FluidAction.EXECUTE) == 250,
+                    "Spell Dispenser rejected a Create regular mana potion fluid");
+            helper.assertTrue(blockEntity.getCurrentMana() == 925,
+                    "Spell Dispenser did not recover mana from Create potion fluid: " + blockEntity.getCurrentMana());
+        });
+    }
+    static void spellDispenserPlacementStartsAtZeroManaAndStoresOwnerProfile(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            var absolutePos = helper.absolutePos(pos);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+            var blockEntity = level.getBlockEntity(absolutePos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var player = createSpellDispenserPlacer(helper, new BlockPos(0, 2, 0), "spell_dispenser_place_mana_test");
+            var magicData = MagicData.getPlayerMagicData(player);
+            helper.assertTrue(magicData != null, "Spell Dispenser placement test could not resolve player mana data");
+            magicData.setMana(1350.75F);
+            var initialMana = magicData.getMana();
+
+            var state = level.getBlockState(absolutePos);
+            ((SpellDispenser) state.getBlock()).setPlacedBy(level, absolutePos, state, player, new ItemStack(ItemRegistry.SPELL_DISPENSER.get()));
+
+            helper.assertTrue(spellDispenser.getCurrentMana() == 0,
+                    "Spell Dispenser placement should now start at zero mana: " + spellDispenser.getCurrentMana());
+            helper.assertTrue(Math.abs(magicData.getMana() - initialMana) < 1.0e-4F,
+                    "Spell Dispenser placement should no longer drain player mana: " + magicData.getMana());
+            helper.assertTrue(spellDispenser.hasOwnerProfile(), "Spell Dispenser placement did not store the owner profile");
+        });
+    }
+
+    static void creativeSpellDispenserPlacementStartsAtZeroManaAndSkipsOwnerProfile(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            var absolutePos = helper.absolutePos(pos);
+            helper.setBlock(pos, BlockRegistry.CREATIVE_SPELL_DISPENSER.get());
+
+            var blockEntity = level.getBlockEntity(absolutePos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Creative Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var player = createSpellDispenserPlacer(helper, new BlockPos(0, 2, 0), "creative_spell_dispenser_place_test");
+            var state = level.getBlockState(absolutePos);
+            ((SpellDispenser) state.getBlock()).setPlacedBy(level, absolutePos, state, player, new ItemStack(ItemRegistry.CREATIVE_SPELL_DISPENSER.get()));
+
+            helper.assertTrue(spellDispenser.getCurrentMana() == 0,
+                    "Creative Spell Dispenser placement should start at zero mana: " + spellDispenser.getCurrentMana());
+            helper.assertFalse(spellDispenser.hasOwnerProfile(), "Creative Spell Dispenser placement stored an owner profile");
+        });
+    }
+
+    static void creativeSpellDispenserIgnoresOwnerProfileNbt(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var ownerProfile = createSpellDispenserOwnerProfile("creative_spell_dispenser_owner_nbt_test");
+            var originalTag = new net.minecraft.nbt.CompoundTag();
+            SpellDispenserBlockEntity.saveOwnerProfile(originalTag, ownerProfile);
+
+            var creative = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.CREATIVE_SPELL_DISPENSER.get().defaultBlockState());
+            creative.loadWithComponents(originalTag, helper.getLevel().registryAccess());
+            helper.assertFalse(creative.hasOwnerProfile(), "Creative Spell Dispenser restored owner profile from NBT");
+
+            creative.setOwnerProfile(ownerProfile);
+            helper.assertFalse(creative.getUpdateTag(helper.getLevel().registryAccess()).hasUUID("OwnerUuid"),
+                    "Creative Spell Dispenser saved owner UUID into update NBT");
+        });
+    }
+
+    static void creativeSpellDispenserUsesCreativeManaConfig(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.CREATIVE_SPELL_DISPENSER.get());
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Creative Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+            spellDispenser.getInventory().setStackInSlot(0, scrollStack.copy());
+            spellDispenser.setCurrentMana(0);
+
+            try (var ignored = ApprenticeCodexServerConfig.useCreativeSpellDispenserConfigOverrideForGameTest(false, 0.1D)) {
+                var castResult = spellDispenser.tryActivate();
+                helper.assertTrue(castResult.succeeded(), "Creative Spell Dispenser consumed mana while creative mana cost was disabled");
+                helper.assertTrue(spellDispenser.getCurrentMana() == 0,
+                        "Creative Spell Dispenser changed mana while creative mana cost was disabled: " + spellDispenser.getCurrentMana());
+            }
+
+            spellDispenser.clearCooldown();
+            try (var ignored = ApprenticeCodexServerConfig.useCreativeSpellDispenserConfigOverrideForGameTest(true, 0.1D)) {
+                var castResult = spellDispenser.tryActivate();
+                helper.assertTrue(!castResult.succeeded(), "Creative Spell Dispenser cast without mana while creative mana cost was enabled");
+                helper.assertTrue(castResult.insufficientMana(), "Creative Spell Dispenser returned the wrong failure with creative mana cost enabled");
+            }
+
+            assertNoSpellDispenserProxy(helper, pos, scrollStack, "Creative Spell Dispenser mana config test left proxy state behind");
+        });
+    }
+
+    static void creativeSpellDispenserAllowsZeroCooldownMultiplier(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.CREATIVE_SPELL_DISPENSER.get());
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Creative Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+            spellDispenser.getInventory().setStackInSlot(0, scrollStack.copy());
+            spellDispenser.setCurrentMana(SpellDispenserManaHelper.MAX_MANA);
+
+            try (var ignored = ApprenticeCodexServerConfig.useCreativeSpellDispenserConfigOverrideForGameTest(true, 0.0D)) {
+                var castResult = spellDispenser.tryActivate();
+                helper.assertTrue(castResult.succeeded(), "Creative Spell Dispenser failed to cast with zero cooldown multiplier");
+                helper.assertTrue(castResult.cooldownTicks() == 0,
+                        "Creative Spell Dispenser did not allow zero cooldown multiplier: " + castResult.cooldownTicks());
+                helper.assertFalse(spellDispenser.isCoolingDown(), "Creative Spell Dispenser entered cooldown with zero cooldown multiplier");
+            }
+
+            assertNoSpellDispenserProxy(helper, pos, scrollStack, "Creative Spell Dispenser cooldown config test left proxy state behind");
+        });
+    }
+
+    static void creativeSpellDispenserDropsNoStoredItems(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            var absolutePos = helper.absolutePos(pos);
+            helper.setBlock(pos, BlockRegistry.CREATIVE_SPELL_DISPENSER.get());
+
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Creative Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get());
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.SPELL_SLOT_INDEX, scrollStack);
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START, new ItemStack(Items.GLASS_BOTTLE));
+
+            level.setBlock(absolutePos, Blocks.AIR.defaultBlockState(), 3);
+            var drops = getFreshItemDrops(level, absolutePos, 1.5D);
+            helper.assertTrue(drops.isEmpty(), "Creative Spell Dispenser dropped stored items: " + drops.size());
+        });
+    }
+
+    static void creativeSpellDispenserRestrictsMenuAccess(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var survivalPlayer = createSpellDispenserPlacer(helper, new BlockPos(0, 2, 0), "creative_spell_dispenser_menu_survival");
+            var creativePlayer = createSpellDispenserPlacer(helper, new BlockPos(0, 2, 0), "creative_spell_dispenser_menu_creative");
+            creativePlayer.gameMode.changeGameModeForPlayer(net.minecraft.world.level.GameType.CREATIVE);
+
+            helper.assertFalse(
+                    SpellDispenserVariant.canUseCreativeVariant(survivalPlayer),
+                    "Creative Spell Dispenser menu allowed a survival player without permissions"
+            );
+            helper.assertTrue(
+                    SpellDispenserVariant.canUseCreativeVariant(creativePlayer),
+                    "Creative Spell Dispenser menu rejected a creative player"
+            );
+        });
+    }
+
+    static void spellDispenserSneakPlacementStartsAtZeroMana(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            var absolutePos = helper.absolutePos(pos);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+            var blockEntity = level.getBlockEntity(absolutePos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var player = createSpellDispenserPlacer(helper, new BlockPos(0, 2, 0), "spell_dispenser_place_sneak_test");
+            player.setShiftKeyDown(true);
+            var magicData = MagicData.getPlayerMagicData(player);
+            helper.assertTrue(magicData != null, "Spell Dispenser sneak placement test could not resolve player mana data");
+            magicData.setMana(640.5F);
+            var initialMana = magicData.getMana();
+
+            var state = level.getBlockState(absolutePos);
+            ((SpellDispenser) state.getBlock()).setPlacedBy(level, absolutePos, state, player, new ItemStack(ItemRegistry.SPELL_DISPENSER.get()));
+
+            helper.assertTrue(spellDispenser.getCurrentMana() == 0,
+                    "Spell Dispenser sneak placement should start at zero mana: " + spellDispenser.getCurrentMana());
+            helper.assertTrue(Math.abs(magicData.getMana() - initialMana) < 1.0e-4F,
+                    "Spell Dispenser sneak placement should not drain player mana: " + magicData.getMana());
+            helper.assertTrue(spellDispenser.hasOwnerProfile(), "Spell Dispenser sneak placement did not store the owner profile");
+        });
+    }
+    static void spellDispenserRefillsFromHighestManaPotionThatFits(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var overflowPotion = createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_THREE.get());
+            var fittingPotion = createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_TWO.get());
+            spellDispenser.setCurrentMana(780);
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START, overflowPotion.copy());
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START + 1, fittingPotion.copy());
+
+            for (var tick = 0; tick < SpellDispenserManaHelper.REFILL_INTERVAL_TICKS; ++tick) {
+                SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            }
+
+            helper.assertTrue(spellDispenser.getCurrentMana() == 930,
+                    "Spell Dispenser did not pick the highest refill that still fits: " + spellDispenser.getCurrentMana());
+            helper.assertTrue(
+                    ItemStack.isSameItemSameComponents(
+                            spellDispenser.getInventory().getStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START),
+                            overflowPotion
+                    ),
+                    "Spell Dispenser consumed a potion that should have overflowed"
+            );
+            helper.assertTrue(
+                    spellDispenser.getInventory().getStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START + 1).is(Items.GLASS_BOTTLE),
+                    "Spell Dispenser did not leave a glass bottle after consuming a direct mana potion"
+            );
+        });
+    }
+    static void spellDispenserRefillsFromGlowEnergyFlask(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var glowEnergyFlask = createFilledSpellcastersFlask(
+                    helper.getLevel().registryAccess(),
+                    createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get()),
+                    1,
+                    1
+            );
+            spellDispenser.setCurrentMana(700);
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START, glowEnergyFlask);
+
+            for (var tick = 0; tick < SpellDispenserManaHelper.REFILL_INTERVAL_TICKS; ++tick) {
+                SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            }
+
+            helper.assertTrue(spellDispenser.getCurrentMana() == 850,
+                    "Spell Dispenser did not apply Glow Energy to flask refill mana: " + spellDispenser.getCurrentMana());
+            helper.assertTrue(
+                    spellDispenser.getInventory().getStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START).is(ItemRegistry.SPELLCASTERS_FLASK.get()),
+                    "Spell Dispenser replaced the flask item instead of emptying one dose"
+            );
+            helper.assertTrue(
+                    !SpellcastersFlask.isFilled(spellDispenser.getInventory().getStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START)),
+                    "Spell Dispenser did not consume exactly one flask dose"
+            );
+        });
+    }
+    static void spellDispenserActivationFailsWhenManaIsInsufficient(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var spell = SpellRegistry.COMPOUND_PHIAL.get();
+            var requiredMana = spell.getManaCost(1);
+            spellDispenser.getInventory().setStackInSlot(0, createSpellScroll(spell));
+            spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_low_mana_test"));
+            spellDispenser.setCurrentMana(Math.max(0, requiredMana - 1));
+
+            var castResult = spellDispenser.tryActivate();
+            helper.assertTrue(!castResult.succeeded(), "Spell Dispenser activated even though it had insufficient mana");
+            helper.assertTrue(castResult.insufficientMana(), "Spell Dispenser returned the wrong failure for insufficient mana");
+            helper.assertTrue(spellDispenser.getCurrentMana() == Math.max(0, requiredMana - 1),
+                    "Spell Dispenser changed mana despite rejecting the cast");
+        });
+    }
+    static void spellDispenserContinuousCastStopsWhenManaRunsOut(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+            helper.setBlock(pos, helper.getBlockState(pos).setValue(SpellDispenser.TRIGGERED, true));
+
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.getSpell(
+                    ResourceLocation.fromNamespaceAndPath("irons_spellbooks", "blaze_storm")
+            );
+            spellDispenser.getInventory().setStackInSlot(0, createSpellScroll(spell));
+            spellDispenser.setOwnerProfile(createSpellDispenserOwnerProfile("spell_dispenser_continuous_low_mana_test"));
+            spellDispenser.setCurrentMana(spell.getManaCost(1));
+
+            var startResult = spellDispenser.tryActivate();
+            helper.assertTrue(startResult.succeeded(), "Spell Dispenser failed to start a CONTINUOUS cast with exactly one cost worth of mana");
+            helper.assertTrue(spellDispenser.hasActiveContinuousCast(), "Spell Dispenser did not keep the CONTINUOUS session active after start");
+            helper.assertTrue(spellDispenser.getCurrentMana() == 0,
+                    "Spell Dispenser did not consume the initial CONTINUOUS mana cost at start: " + spellDispenser.getCurrentMana());
+
+            for (var tick = 0; tick < 40 && spellDispenser.hasActiveContinuousCast(); ++tick) {
+                SpellDispenserBlockEntity.serverTick(level, pos, helper.getBlockState(pos), spellDispenser);
+            }
+
+            helper.assertFalse(spellDispenser.hasActiveContinuousCast(),
+                    "Spell Dispenser kept a CONTINUOUS cast active after its mana ran out");
+            helper.assertTrue(spellDispenser.getCurrentMana() == 0,
+                    "Spell Dispenser mana changed unexpectedly after a mana-depleted CONTINUOUS stop");
+        });
+    }
+    static void spellDispenserCreateContinuousCastRequiresDisableBeforeRestart(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.getSpell(
+                    ResourceLocation.fromNamespaceAndPath("irons_spellbooks", "blaze_storm")
+            );
+            var scrollStack = createSpellScroll(spell);
+            var mountedInventory = new ItemStackHandler(1);
+            mountedInventory.setStackInSlot(0, scrollStack.copy());
+            var harness = createSpellDispenserMovementHarness(level, castPos, mountedInventory, createSpellDispenserOwnerProfile("spell_dispenser_create_reset_test"));
+
+            startCreateSpellDispenserMovement(harness);
+            tickCreateSpellDispenserMovement(harness);
+            helper.assertTrue(hasCreateSpellDispenserContinuousCast(harness),
+                    "Create-mounted Spell Dispenser did not start CONTINUOUS casting when enabled");
+
+            var maxTicks = spell.getEffectiveCastTime(1, new FakePlayer(level, createSpellDispenserOwnerProfile("spell_dispenser_create_reset_probe")));
+            for (var tick = 0; tick <= maxTicks + 40 && hasCreateSpellDispenserContinuousCast(harness); tick++) {
+                tickCreateSpellDispenserMovement(harness);
+            }
+
+            helper.assertFalse(hasCreateSpellDispenserContinuousCast(harness),
+                    "Create-mounted Spell Dispenser CONTINUOUS cast did not stop at its duration cap");
+            helper.assertTrue(createSpellDispenserRequiresReset(harness),
+                    "Create-mounted Spell Dispenser did not enter reset-required state after held completion");
+            helper.assertTrue(createSpellDispenserIsCoolingDown(harness),
+                    "Create-mounted Spell Dispenser did not enter cooldown after held completion");
+
+            tickCreateSpellDispenserMovement(harness);
+            helper.assertFalse(hasCreateSpellDispenserContinuousCast(harness),
+                    "Create-mounted Spell Dispenser restarted CONTINUOUS casting without being disabled first");
+
+            setCreateSpellDispenserDisabled(harness, true);
+            tickCreateSpellDispenserMovement(harness);
+            helper.assertFalse(createSpellDispenserRequiresReset(harness),
+                    "Create-mounted Spell Dispenser did not clear reset-required state when disabled");
+
+            for (var tick = 0; tick < spell.getSpellCooldown() && createSpellDispenserIsCoolingDown(harness); tick++) {
+                tickCreateSpellDispenserMovement(harness);
+            }
+            helper.assertFalse(createSpellDispenserIsCoolingDown(harness),
+                    "Create-mounted Spell Dispenser cooldown did not expire after the expected ticks");
+
+            setCreateSpellDispenserDisabled(harness, false);
+            tickCreateSpellDispenserMovement(harness);
+            helper.assertTrue(hasCreateSpellDispenserContinuousCast(harness),
+                    "Create-mounted Spell Dispenser did not restart CONTINUOUS casting after disable/enable");
+
+            stopCreateSpellDispenserMovement(harness);
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Create-mounted Spell Dispenser left proxy state behind after stopMoving");
+        });
+    }
+    static void spellDispenserCreateContinuousCastStopsWhenDisabled(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.FIRE_BREATH_SPELL.get();
+            var scrollStack = createSpellScroll(spell);
+            var mountedInventory = new ItemStackHandler(1);
+            mountedInventory.setStackInSlot(0, scrollStack.copy());
+            var harness = createSpellDispenserMovementHarness(level, castPos, mountedInventory, createSpellDispenserOwnerProfile("spell_dispenser_create_disable_test"));
+
+            startCreateSpellDispenserMovement(harness);
+            tickCreateSpellDispenserMovement(harness);
+            helper.assertTrue(hasCreateSpellDispenserContinuousCast(harness),
+                    "Create-mounted Spell Dispenser did not start Fire Breath CONTINUOUS casting");
+
+            for (var tick = 0; tick < 20; tick++) {
+                tickCreateSpellDispenserMovement(harness);
+            }
+
+            setCreateSpellDispenserDisabled(harness, true);
+            tickCreateSpellDispenserMovement(harness);
+            helper.assertFalse(hasCreateSpellDispenserContinuousCast(harness),
+                    "Create-mounted Spell Dispenser kept a CONTINUOUS session active after disable");
+            helper.assertFalse(createSpellDispenserRequiresReset(harness),
+                    "Create-mounted Spell Dispenser incorrectly required reset after disable cancellation");
+            helper.assertTrue(createSpellDispenserIsCoolingDown(harness),
+                    "Create-mounted Spell Dispenser did not enter cooldown after cancelling an already-casting CONTINUOUS spell");
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Create-mounted Spell Dispenser left proxy state behind after disable");
+        });
+    }
+    static void spellDispenserCreateInstantCastUsesCooldown(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var spell = SpellRegistry.COMPOUND_PHIAL.get();
+            var scrollStack = createSpellScroll(spell);
+            var mountedInventory = new ItemStackHandler(1);
+            mountedInventory.setStackInSlot(0, scrollStack.copy());
+            var harness = createSpellDispenserMovementHarness(level, castPos, mountedInventory, createSpellDispenserOwnerProfile("spell_dispenser_create_instant_test"));
+
+            startCreateSpellDispenserMovement(harness);
+            visitCreateSpellDispenserPosition(harness, castPos);
+            helper.assertTrue(createSpellDispenserIsCoolingDown(harness),
+                    "Create-mounted Spell Dispenser did not enter cooldown after a non-continuous cast");
+
+            var projectileBox = new AABB(castPos).inflate(5.0D);
+            var firstProjectileCount = level.getEntitiesOfClass(CompoundPhialProjectileEntity.class, projectileBox).size();
+            helper.assertTrue(firstProjectileCount > 0,
+                    "Create-mounted Spell Dispenser did not spawn a projectile on the first cast");
+
+            visitCreateSpellDispenserPosition(harness, castPos);
+            var secondProjectileCount = level.getEntitiesOfClass(CompoundPhialProjectileEntity.class, projectileBox).size();
+            helper.assertTrue(secondProjectileCount == firstProjectileCount,
+                    "Create-mounted Spell Dispenser ignored cooldown and fired again immediately");
+
+            var expectedCooldown = spell.getSpellCooldown()
+                    + spell.getEffectiveCastTime(1, new FakePlayer(level, createSpellDispenserOwnerProfile("spell_dispenser_create_instant_probe")));
+            for (var tick = 0; tick < expectedCooldown && createSpellDispenserIsCoolingDown(harness); tick++) {
+                tickCreateSpellDispenserMovement(harness);
+            }
+
+            helper.assertFalse(createSpellDispenserIsCoolingDown(harness),
+                    "Create-mounted Spell Dispenser cooldown did not expire after the expected ticks");
+            visitCreateSpellDispenserPosition(harness, castPos);
+            var thirdProjectileCount = level.getEntitiesOfClass(CompoundPhialProjectileEntity.class, projectileBox).size();
+            helper.assertTrue(thirdProjectileCount > secondProjectileCount,
+                    "Create-mounted Spell Dispenser did not fire again after cooldown expired");
+        });
+    }
+    static void spellDispenserCreateAppliesCooldownMultiplier(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            try (var ignored = ApprenticeCodexServerConfig.useSpellDispenserConfigOverrideForGameTest(
+                    true,
+                    false,
+                    List.of(),
+                    0.1D
+            )) {
+                var level = (ServerLevel) helper.getLevel();
+                var castPos = new BlockPos(0, 1, 0);
+                var spell = io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get();
+                var scrollStack = createSpellScroll(spell);
+                var mountedInventory = new ItemStackHandler(1);
+                mountedInventory.setStackInSlot(0, scrollStack.copy());
+                var harness = createSpellDispenserMovementHarness(level, castPos, mountedInventory, null);
+
+                startCreateSpellDispenserMovement(harness);
+                visitCreateSpellDispenserPosition(harness, castPos);
+                helper.assertTrue(createSpellDispenserIsCoolingDown(harness),
+                        "Create-mounted Spell Dispenser did not enter cooldown after a scaled-cooldown cast");
+
+                var expectedCooldown = Math.max(1, (int) Math.ceil(spell.getSpellCooldown() * 0.1D));
+                for (var tick = 0; tick < expectedCooldown && createSpellDispenserIsCoolingDown(harness); tick++) {
+                    tickCreateSpellDispenserMovement(harness);
+                }
+
+                helper.assertFalse(createSpellDispenserIsCoolingDown(harness),
+                        "Create-mounted Spell Dispenser cooldown did not expire after the scaled cooldown ticks");
+                assertNoSpellDispenserProxy(helper, castPos, scrollStack,
+                        "Create-mounted Spell Dispenser cooldown multiplier test left proxy state behind");
+            }
+        });
+    }
+    static void spellDispenserCreateAllowsOwnerOptionalSpellWithoutOwnerProfile(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get());
+            var mountedInventory = new ItemStackHandler(1);
+            mountedInventory.setStackInSlot(0, scrollStack.copy());
+            var harness = createSpellDispenserMovementHarness(level, castPos, mountedInventory, null);
+
+            startCreateSpellDispenserMovement(harness);
+            visitCreateSpellDispenserPosition(harness, castPos);
+            helper.assertTrue(createSpellDispenserIsCoolingDown(harness),
+                    "Create-mounted Spell Dispenser did not cast an owner-optional spell without an owner profile");
+            assertNoSpellDispenserProxy(helper, castPos, scrollStack, "Create-mounted Spell Dispenser owner-optional cast left proxy state behind");
+        });
+    }
+    static void spellDispenserCreateRefillsFromFuelStorageAndReturnsBottle(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var mountedInventory = new ItemStackHandler(SpellDispenserBlockEntity.INVENTORY_SLOT_COUNT);
+            var externalInventory = new ItemStackHandler(2);
+            externalInventory.setStackInSlot(0,
+                    createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get()));
+
+            var harness = createSpellDispenserMovementHarness(
+                    level,
+                    castPos,
+                    mountedInventory,
+                    createSpellDispenserOwnerProfile("spell_dispenser_create_fuel_refill_test"),
+                    800,
+                    externalInventory,
+                    true
+            );
+
+            startCreateSpellDispenserMovement(harness);
+            for (var tick = 0; tick < SpellDispenserManaHelper.REFILL_INTERVAL_TICKS; ++tick) {
+                tickCreateSpellDispenserMovement(harness);
+            }
+
+            helper.assertTrue(externalInventory.getStackInSlot(0).is(Items.GLASS_BOTTLE),
+                    "Create-mounted Spell Dispenser did not return a glass bottle to accessible contraption storage");
+        });
+    }
+    static void spellDispenserCreateIgnoresProtectedFuelStorage(GameTestHelper helper) {
+        if (skipWhenCreateMissing(helper)) {
+            return;
+        }
+
+        helper.succeedIf(() -> {
+            var level = (ServerLevel) helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var mountedInventory = new ItemStackHandler(SpellDispenserBlockEntity.INVENTORY_SLOT_COUNT);
+            var externalInventory = new ItemStackHandler(2);
+            var manaPotion = createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get());
+            externalInventory.setStackInSlot(0, manaPotion.copy());
+
+            var harness = createSpellDispenserMovementHarness(
+                    level,
+                    castPos,
+                    mountedInventory,
+                    createSpellDispenserOwnerProfile("spell_dispenser_create_protected_storage_test"),
+                    800,
+                    externalInventory,
+                    false
+            );
+
+            startCreateSpellDispenserMovement(harness);
+            for (var tick = 0; tick < SpellDispenserManaHelper.REFILL_INTERVAL_TICKS; ++tick) {
+                tickCreateSpellDispenserMovement(harness);
+            }
+
+            helper.assertTrue(ItemStack.isSameItemSameComponents(externalInventory.getStackInSlot(0), manaPotion),
+                    "Create-mounted Spell Dispenser consumed a potion from protected contraption storage");
+        });
+    }
+    static void spellDispenserBlockEntityRejectsActivationWithoutOwnerProfile(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            spellDispenser.getInventory().setStackInSlot(0, createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.SPECTRAL_HAMMER_SPELL.get()));
+
+            var castResult = spellDispenser.tryActivate();
+            helper.assertTrue(!castResult.succeeded(), "Spell Dispenser activated without an owner profile");
+            helper.assertTrue(castResult.missingOwnerProfile(), "Spell Dispenser returned the wrong failure for missing owner profile");
+        });
+    }
+    static void spellDispenserBlockEntityAllowsOwnerOptionalActivationWithoutOwnerProfile(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get());
+            spellDispenser.getInventory().setStackInSlot(0, scrollStack);
+
+            var castResult = spellDispenser.tryActivate();
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser rejected an owner-optional spell without an owner profile");
+            assertNoSpellDispenserProxy(helper, pos, scrollStack, "Spell Dispenser block entity owner-optional cast left proxy state behind");
+        });
+    }
+    static void spellDispenserSidedAutomationProtectsScrollSlot(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            var itemHandler = helper.getLevel().getCapability(
+                    net.neoforged.neoforge.capabilities.Capabilities.ItemHandler.BLOCK,
+                    helper.absolutePos(pos),
+                    Direction.UP
+            );
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.HEAL_SPELL.get());
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.SPELL_SLOT_INDEX, scrollStack.copy());
+
+            helper.assertTrue(itemHandler != null, "Spell Dispenser item capability was not exposed");
+            helper.assertTrue(itemHandler != null && !itemHandler.isItemValid(0, scrollStack),
+                    "Spell Dispenser sided automation exposed the scroll slot as insertable");
+
+            var remainder = itemHandler == null ? scrollStack.copy() : itemHandler.insertItem(0, scrollStack.copy(), false);
+            helper.assertTrue(ItemStack.isSameItemSameComponents(remainder, scrollStack),
+                    "Spell Dispenser sided automation consumed a scroll while rejecting the protected slot");
+            helper.assertTrue(itemHandler != null && ItemStack.isSameItemSameComponents(itemHandler.getStackInSlot(0), scrollStack),
+                    "Spell Dispenser sided automation inserted a scroll into the protected slot");
+            helper.assertTrue(itemHandler != null && itemHandler.extractItem(0, 1, false).isEmpty(),
+                    "Spell Dispenser sided automation extracted the protected scroll slot");
+        });
+    }
+    static void spellDispenserSidedAutomationExtractsOnlyEmptyContainers(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var pos = new BlockPos(0, 1, 0);
+            helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+            var blockEntity = helper.getBlockEntity(pos);
+            helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity, "Spell Dispenser block entity was not created");
+            var spellDispenser = (SpellDispenserBlockEntity) blockEntity;
+            var itemHandler = helper.getLevel().getCapability(
+                    net.neoforged.neoforge.capabilities.Capabilities.ItemHandler.BLOCK,
+                    helper.absolutePos(pos),
+                    Direction.UP
+            );
+            var filledFlask = createFilledSpellcastersFlask(
+                    helper.getLevel().registryAccess(),
+                    createInstantManaPotion(io.redspace.ironsspellbooks.registries.PotionRegistry.INSTANT_MANA_ONE.get()),
+                    1,
+                    0
+            );
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START, filledFlask);
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START + 1, new ItemStack(ItemRegistry.SPELLCASTERS_FLASK.get()));
+            spellDispenser.getInventory().setStackInSlot(SpellDispenserBlockEntity.FLASK_SLOT_START + 2, new ItemStack(Items.GLASS_BOTTLE));
+
+            helper.assertTrue(itemHandler != null, "Spell Dispenser sided automation capability was not exposed");
+            helper.assertTrue(itemHandler != null && itemHandler.extractItem(SpellDispenserBlockEntity.FLASK_SLOT_START, 1, false).isEmpty(),
+                    "Spell Dispenser sided automation extracted a filled mana container");
+            helper.assertTrue(itemHandler != null
+                            && itemHandler.extractItem(SpellDispenserBlockEntity.FLASK_SLOT_START + 1, 1, false).is(ItemRegistry.SPELLCASTERS_FLASK.get()),
+                    "Spell Dispenser sided automation failed to extract an empty flask");
+            helper.assertTrue(itemHandler != null
+                            && itemHandler.extractItem(SpellDispenserBlockEntity.FLASK_SLOT_START + 2, 1, false).is(Items.GLASS_BOTTLE),
+                    "Spell Dispenser sided automation failed to extract a glass bottle");
+        });
+    }
+    static void spellDispenserMountedMenuKeepsContraptionInventoryAccessible(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var player = new FakePlayer(helper.getLevel(), new GameProfile(UUID.randomUUID(), "spell_dispenser_mounted_menu_test"));
+            var mountedInventory = new ItemStackHandler(SpellDispenserBlockEntity.INVENTORY_SLOT_COUNT);
+            var scrollStack = createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get());
+            mountedInventory.setStackInSlot(0, scrollStack.copy());
+            var currentMana = new AtomicInteger(320);
+            var ownerName = "spell_dispenser_mounted_menu_owner_test";
+
+            var menu = SpellDispenserMenu.createMounted(0, new Inventory(player), BlockPos.ZERO, mountedInventory, true, ownerName, currentMana::get);
+            helper.assertTrue(menu.stillValid(player), "Spell Dispenser mounted menu closed because it expected a world block entity");
+            helper.assertTrue(ItemStack.isSameItemSameComponents(menu.getSpellSource(), scrollStack),
+                    "Spell Dispenser mounted menu did not expose the mounted inventory stack");
+            helper.assertTrue(menu.isReadyToCast(player),
+                    "Spell Dispenser mounted menu did not report a valid mounted scroll as ready");
+            helper.assertTrue(menu.getCurrentMana() == 320,
+                    "Spell Dispenser mounted menu did not expose the mounted mana value");
+            helper.assertTrue(ownerName.equals(menu.getOwnerName()),
+                    "Spell Dispenser mounted menu did not expose the mounted owner name");
+
+            currentMana.set(180);
+            helper.assertTrue(menu.getCurrentMana() == 180,
+                    "Spell Dispenser mounted menu did not follow the live contraption mana source");
+
+            var clientMenu = SpellDispenserMenu.createMounted(1, new Inventory(player), BlockPos.ZERO, mountedInventory, true, ownerName, 320);
+            clientMenu.setData(0, 180);
+            helper.assertTrue(clientMenu.getCurrentMana() == 180,
+                    "Spell Dispenser mounted menu did not accept mana updates from menu data sync");
+            helper.assertTrue(ownerName.equals(clientMenu.getOwnerName()),
+                    "Spell Dispenser mounted client menu did not retain the synced owner name");
+        });
+    }
+    static void spellDispenserMountedMenuTreatsOwnerOptionalSpellAsReadyWithoutOwnerProfile(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var player = new FakePlayer(helper.getLevel(), new GameProfile(UUID.randomUUID(), "spell_dispenser_owner_optional_menu_test"));
+            var mountedInventory = new ItemStackHandler(SpellDispenserBlockEntity.INVENTORY_SLOT_COUNT);
+            mountedInventory.setStackInSlot(0, createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get()));
+
+            var menu = SpellDispenserMenu.createMounted(0, new Inventory(player), BlockPos.ZERO, mountedInventory, false, null, 320);
+            helper.assertTrue(menu.isReadyToCast(player),
+                    "Spell Dispenser mounted menu still required an owner for an owner-optional spell");
+        });
+    }
+    static void spellDispenserIsTaggedForContraptionControls(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var spellDispenserStack = new ItemStack(ItemRegistry.SPELL_DISPENSER.get());
+            helper.assertTrue(spellDispenserStack.is(CREATE_CONTRAPTION_CONTROLLED),
+                    "Spell Dispenser is missing create:contraption_controlled and cannot be selected by Contraption Controls");
+        });
+    }
+
+    static void creativeSpellDispenserIsTaggedForContraptionControls(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var spellDispenserStack = new ItemStack(ItemRegistry.CREATIVE_SPELL_DISPENSER.get());
+            helper.assertTrue(spellDispenserStack.is(CREATE_CONTRAPTION_CONTROLLED),
+                    "Creative Spell Dispenser is missing create:contraption_controlled and cannot be selected by Contraption Controls");
+        });
+    }
+
+    static void spellDispenserOwnerProfileCanBeReadFromSavedTag(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var ownerProfile = createSpellDispenserOwnerProfile("spell_dispenser_owner_tag_test");
+            var blockEntity = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            blockEntity.setOwnerProfile(ownerProfile);
+
+            var restoredOwner = SpellDispenserBlockEntity.readOwnerProfile(blockEntity.getUpdateTag(helper.getLevel().registryAccess()));
+            helper.assertTrue(restoredOwner != null, "Spell Dispenser owner profile helper returned null");
+            helper.assertTrue(restoredOwner != null && ownerProfile.getId().equals(restoredOwner.getId()),
+                    "Spell Dispenser owner UUID helper returned the wrong value");
+            helper.assertTrue(restoredOwner != null && ownerProfile.getName().equals(restoredOwner.getName()),
+                    "Spell Dispenser owner name helper returned the wrong value");
+        });
+    }
+    static void spellDispenserOwnerProfilePersistsThroughNbt(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var ownerProfile = createSpellDispenserOwnerProfile("spell_dispenser_owner_save_test");
+            var original = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            original.setOwnerProfile(ownerProfile);
+
+            var restored = new SpellDispenserBlockEntity(BlockPos.ZERO, BlockRegistry.SPELL_DISPENSER.get().defaultBlockState());
+            restored.loadWithComponents(original.getUpdateTag(helper.getLevel().registryAccess()), helper.getLevel().registryAccess());
+
+            helper.assertTrue(restored.hasOwnerProfile(), "Spell Dispenser owner profile was not restored from NBT");
+            helper.assertTrue(restored.getOwnerProfile() != null && ownerProfile.getId().equals(restored.getOwnerProfile().getId()),
+                    "Spell Dispenser owner UUID changed during NBT round-trip");
+            helper.assertTrue(restored.getOwnerProfile() != null && ownerProfile.getName().equals(restored.getOwnerProfile().getName()),
+                    "Spell Dispenser owner name changed during NBT round-trip");
+        });
+    }
+    static void spellDispenserCastHelperSupportsDiagonalVectorFacing(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = helper.getLevel();
+            var castPos = new BlockPos(0, 1, 0);
+            var scrollStack = createSpellScroll(SpellRegistry.COMPOUND_PHIAL.get());
+
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    (ServerLevel) level,
+                    castPos,
+                    new Vec3(1.0D, 0.0D, 1.0D),
+                    scrollStack,
+                    createSpellDispenserOwnerProfile("spell_dispenser_diagonal_vector_test")
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser cast helper failed to cast from a diagonal forward vector");
+
+            var projectileBox = new AABB(castPos).inflate(5.0D);
+            var projectiles = level.getEntitiesOfClass(CompoundPhialProjectileEntity.class, projectileBox);
+            helper.assertTrue(!projectiles.isEmpty(), "Spell Dispenser diagonal vector cast did not spawn a Compound Phial projectile");
+
+            var projectile = projectiles.get(0);
+            var motion = projectile.getDeltaMovement();
+            helper.assertTrue(Math.abs(motion.x) > 0.01D, "Spell Dispenser diagonal vector cast kept the projectile X motion at zero");
+            helper.assertTrue(Math.abs(motion.z) > 0.01D, "Spell Dispenser diagonal vector cast kept the projectile Z motion at zero");
+            helper.assertTrue(Math.signum(motion.x) == Math.signum(motion.z),
+                    "Spell Dispenser diagonal vector cast did not preserve the intended diagonal quadrant: " + motion);
+        });
+    }
+    static void spellDispenserCastHelperSupportsSpectralHammer(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            var level = helper.getLevel();
+            var relativeCastPos = new BlockPos(0, 1, 0);
+            var relativeTargetPos = new BlockPos(0, 1, 3);
+            var castPos = helper.absolutePos(relativeCastPos);
+            helper.setBlock(relativeTargetPos, Blocks.STONE);
+
+            var castResult = SpellDispenserCastHelper.tryCast(
+                    (ServerLevel) level,
+                    castPos,
+                    Direction.SOUTH,
+                    createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.SPECTRAL_HAMMER_SPELL.get()),
+                    createSpellDispenserOwnerProfile("spell_dispenser_spectral_hammer_test")
+            );
+            helper.assertTrue(castResult.succeeded(), "Spell Dispenser cast helper failed to start Spectral Hammer");
+
+            var hammerBox = new AABB(castPos).inflate(8.0D);
+            var hammers = level.getEntitiesOfClass(SpectralHammer.class, hammerBox);
+            helper.assertTrue(!hammers.isEmpty(), "Spell Dispenser cast did not spawn Spectral Hammer");
+
+            var hammer = hammers.get(0);
+            for (var tick = 0; tick < 20 && !hammer.isRemoved(); tick++) {
+                hammer.tick();
+            }
+
+            helper.assertTrue(true, "Spectral Hammer ticked without crashing");
+        });
+    }
+
+    static void spellDispenserComparatorOutputMatchesStoredMana(GameTestHelper helper) {
+        helper.succeedIf(() -> {
+            assertSpellDispenserComparatorOutput(helper, new BlockPos(0, 1, 0), 0, false, 0);
+            assertSpellDispenserComparatorOutput(helper, new BlockPos(1, 1, 0), 1, false, 1);
+            assertSpellDispenserComparatorOutput(helper, new BlockPos(2, 1, 0), SpellDispenserManaHelper.MAX_MANA / 2, false, 8);
+            assertSpellDispenserComparatorOutput(helper, new BlockPos(3, 1, 0), SpellDispenserManaHelper.MAX_MANA, false, 15);
+            assertSpellDispenserComparatorOutput(helper, new BlockPos(4, 1, 0), 0, true, 0);
+        });
+    }
+
+    private static FakePlayer createSpellDispenserPlacer(GameTestHelper helper, BlockPos pos, String profileName) {
+        var player = new FakePlayer(helper.getLevel(), new GameProfile(UUID.randomUUID(), profileName));
+        player.gameMode.changeGameModeForPlayer(net.minecraft.world.level.GameType.SURVIVAL);
+        var absolutePos = helper.absoluteVec(Vec3.atBottomCenterOf(pos));
+        player.setPos(absolutePos.x, absolutePos.y, absolutePos.z);
+        return player;
+    }
+
+    private static void assertSpellDispenserComparatorOutput(
+            GameTestHelper helper,
+            BlockPos pos,
+            int currentMana,
+            boolean insertInventoryScroll,
+            int expectedOutput
+    ) {
+        helper.setBlock(pos, BlockRegistry.SPELL_DISPENSER.get());
+
+        var blockEntity = helper.getBlockEntity(pos);
+        helper.assertTrue(blockEntity instanceof SpellDispenserBlockEntity,
+                "Spell Dispenser block entity was not created");
+
+        if (blockEntity instanceof SpellDispenserBlockEntity spellDispenser) {
+            spellDispenser.setCurrentMana(currentMana);
+            if (insertInventoryScroll) {
+                spellDispenser.getInventory().setStackInSlot(
+                        SpellDispenserBlockEntity.SPELL_SLOT_INDEX,
+                        createSpellScroll(io.redspace.ironsspellbooks.api.registry.SpellRegistry.MAGIC_MISSILE_SPELL.get())
+                );
+            }
+        }
+
+        var absolutePos = helper.absolutePos(pos);
+        var state = helper.getLevel().getBlockState(absolutePos);
+        helper.assertTrue(state.hasAnalogOutputSignal(),
+                "Spell Dispenser should advertise comparator output");
+
+        var output = state.getAnalogOutputSignal(helper.getLevel(), absolutePos);
+        helper.assertTrue(output == expectedOutput,
+                "Spell Dispenser comparator output mismatch: expected " + expectedOutput + " but got " + output);
+    }
+
+    private static ItemStack createSpellScroll(AbstractSpell spell) {
+        var stack = new ItemStack(io.redspace.ironsspellbooks.registries.ItemRegistry.SCROLL.get());
+        ISpellContainer.createScrollContainer(spell, 1, stack);
+        return stack;
+    }
+
+    private static List<ItemEntity> getFreshItemDrops(ServerLevel level, BlockPos pos, double radius) {
+        return level.getEntitiesOfClass(
+                ItemEntity.class,
+                new AABB(pos).inflate(radius),
+                itemEntity -> itemEntity.getAge() <= 1
+        );
+    }
+
+    private static ItemStack createInstantManaPotion(net.minecraft.world.item.alchemy.Potion potion) {
+        return PotionContentsHelper.createPotionStack(Items.POTION, potion);
+    }
+
+    private static FluidStack createIronsManaPotionFluid(net.minecraft.world.item.alchemy.Potion potion, int amountMb) {
+        var fluid = io.redspace.ironsspellbooks.fluids.PotionFluid.from(PotionContentsHelper.createPotionStack(Items.POTION, potion));
+        fluid.setAmount(amountMb);
+        return fluid;
+    }
+
+    private static FluidStack createCreateManaPotionFluid(net.minecraft.world.item.alchemy.Potion potion, int amountMb) {
+        var createPotion = BuiltInRegistries.FLUID.get(ResourceLocation.fromNamespaceAndPath("create", "potion"));
+        if (createPotion == null) {
+            return FluidStack.EMPTY;
+        }
+
+        var fluid = new FluidStack(createPotion, amountMb);
+        fluid.set(DataComponents.POTION_CONTENTS, PotionContentsHelper.createPotionStack(Items.POTION, potion).get(DataComponents.POTION_CONTENTS));
+        return fluid;
+    }
+
+    private static ItemStack createFilledSpellcastersFlask(
+            RegistryAccess registryAccess,
+            ItemStack storedItem,
+            int doseCount,
+            int glowEnergyLevel
+    ) {
+        var flask = new ItemStack(ItemRegistry.SPELLCASTERS_FLASK.get());
+        if (glowEnergyLevel > 0) {
+            registryAccess.lookupOrThrow(Registries.ENCHANTMENT)
+                    .get(Enchantments.GLOW_ENERGY)
+                    .ifPresent(enchantment -> flask.enchant(enchantment, glowEnergyLevel));
+        }
+        return SpellcastersFlask.copyWithAddedDoses(flask, storedItem, doseCount);
+    }
+
+    private static ItemStack createFilledAlchemistsFlask(
+            RegistryAccess registryAccess,
+            ItemStack storedItem,
+            int doseCount,
+            int glowEnergyLevel
+    ) {
+        var flask = new ItemStack(ItemRegistry.ALCHEMISTS_FLASK.get());
+        if (glowEnergyLevel > 0) {
+            registryAccess.lookupOrThrow(Registries.ENCHANTMENT)
+                    .get(Enchantments.GLOW_ENERGY)
+                    .ifPresent(enchantment -> flask.enchant(enchantment, glowEnergyLevel));
+        }
+        return SpellcastersFlask.copyWithAddedDoses(flask, storedItem, doseCount);
+    }
+
+    private static SpellDispenserSpellValidator.ValidationResult createSpellDispenserValidation(ItemStack stack, AbstractSpell spell) {
+        return new SpellDispenserSpellValidator.ValidationResult(stack, new SpellData(spell, 1), SpellDispenserSpellValidator.FailureReason.NONE);
+    }
+
+    private static GameProfile createSpellDispenserOwnerProfile(String name) {
+        return new GameProfile(UUID.randomUUID(), name);
+    }
+
+    private static boolean skipWhenCreateMissing(GameTestHelper helper) {
+        if (ModList.get().isLoaded("create")) {
+            return false;
+        }
+
+        // optional 依存の absent 環境では Create 専用テストを成功扱いで抜け、通常検証の起動性を優先する。
+        helper.succeed();
+        return true;
+    }
+
+    private static void assertNoSpellDispenserProxy(GameTestHelper helper, BlockPos castPos, ItemStack spellSource, String message) {
+        var proxyBox = new AABB(castPos).inflate(3.0D);
+        var remainingProxies = helper.getLevel().getEntitiesOfClass(ArmorStand.class, proxyBox, stand ->
+                stand.isInvisible() && stand.getMainHandItem().is(spellSource.getItem()));
+        helper.assertTrue(remainingProxies.isEmpty(), message + ": " + remainingProxies.size());
+
+        var remainingAnchors = helper.getLevel().getEntitiesOfClass(SpellDispenserAnchorEntity.class, proxyBox);
+        helper.assertTrue(remainingAnchors.isEmpty(), message + " (tracked anchors): " + remainingAnchors.size());
+    }
+
+    private static Object createSpellDispenserMovementHarness(
+            ServerLevel level,
+            BlockPos worldPos,
+            ItemStackHandler mountedInventory,
+            GameProfile ownerProfile,
+            int currentMana,
+            ItemStackHandler externalInventory,
+            boolean externalAvailableForFuel
+    ) {
+        return invokeCreateGameTestHook(
+                "createSpellDispenserMovementHarness",
+                new Class<?>[]{ServerLevel.class, BlockPos.class, ItemStackHandler.class, GameProfile.class, int.class, ItemStackHandler.class, boolean.class},
+                level, worldPos, mountedInventory, ownerProfile, currentMana, externalInventory, externalAvailableForFuel
+        );
+    }
+
+    private static Object createSpellDispenserMovementHarness(
+            ServerLevel level,
+            BlockPos worldPos,
+            ItemStackHandler mountedInventory,
+            GameProfile ownerProfile,
+            int currentMana,
+            ItemStackHandler externalInventory,
+            boolean externalAvailableForFuel
+    ) {
+        return invokeCreateGameTestHook(
+                "createSpellDispenserMovementHarness",
+                new Class<?>[]{ServerLevel.class, BlockPos.class, ItemStackHandler.class, GameProfile.class, int.class, ItemStackHandler.class, boolean.class},
+                level, worldPos, mountedInventory, ownerProfile, currentMana, externalInventory, externalAvailableForFuel
+        );
+    }
+
+    private static void startCreateSpellDispenserMovement(Object harness) {
+        invokeCreateGameTestHook("startMoving", new Class<?>[]{Object.class}, harness);
+    }
+
+    private static void tickCreateSpellDispenserMovement(Object harness) {
+        invokeCreateGameTestHook("tick", new Class<?>[]{Object.class}, harness);
+    }
+
+    private static void visitCreateSpellDispenserPosition(Object harness, BlockPos pos) {
+        invokeCreateGameTestHook("visitNewPosition", new Class<?>[]{Object.class, BlockPos.class}, harness, pos);
+    }
+
+    private static void stopCreateSpellDispenserMovement(Object harness) {
+        invokeCreateGameTestHook("stopMoving", new Class<?>[]{Object.class}, harness);
+    }
+
+    private static void setCreateSpellDispenserDisabled(Object harness, boolean disabled) {
+        invokeCreateGameTestHook("setDisabled", new Class<?>[]{Object.class, boolean.class}, harness, disabled);
+    }
+
+    private static boolean hasCreateSpellDispenserContinuousCast(Object harness) {
+        return invokeCreateGameTestHookBoolean("hasRunningContinuousCast", harness);
+    }
+
+    private static boolean createSpellDispenserRequiresReset(Object harness) {
+        return invokeCreateGameTestHookBoolean("requiresContinuousReset", harness);
+    }
+
+    private static boolean createSpellDispenserIsCoolingDown(Object harness) {
+        return invokeCreateGameTestHookBoolean("isCoolingDown", harness);
+    }
+
+    private static boolean invokeCreateGameTestHookBoolean(String methodName, Object harness) {
+        return (boolean) invokeCreateGameTestHook(methodName, new Class<?>[]{Object.class}, harness);
+    }
+
+    private static Object invokeCreateGameTestHook(String methodName, Class<?>[] parameterTypes, Object... args) {
+        try {
+            var hooksClass = Class.forName(CREATE_GAMETEST_HOOKS_CLASS);
+            return hooksClass.getMethod(methodName, parameterTypes).invoke(null, args);
+        } catch (ReflectiveOperationException exception) {
+            throw new IllegalStateException("Create GameTest helper call failed: " + methodName, exception);
+        }
+    }
+
+    private static ItemStack createInitializedPresetStack(Item item) {
+        var stack = new ItemStack(item);
+        if (item instanceof IPresetSpellContainer presetSpellContainer) {
+            presetSpellContainer.initializeSpellContainer(stack);
+        }
+        return stack;
+    }
+}
